% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_getLab.R
\name{anchor_getLab}
\alias{anchor_getLab}
\title{anchor_getLab}
\usage{
anchor_getLab(folder, test, run)
}
\arguments{
\item{folder}{Folder to extract anchor dataframe, either 'input' or 'calibration'.}

\item{test}{Name of test in 'calibration' folder.}

\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}
}
\description{
This function puts dataframe read from anchor file to dataframe of columns of item
label, delta, and step number (if existent).
}
