% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_shift.R
\name{anchor_shift}
\alias{anchor_shift}
\title{anchor_shift}
\usage{
anchor_shift(test, shift)
}
\arguments{
\item{test}{Name of test.}

\item{shift}{Mean of non-DIF anchor deltas in Test A minus that in Test B.}
}
\description{
This function adds anchor shift to item deltas in Test B (in 'output' folder),
keeps step estimates (if any) intact, and outputs shifted deltas in 'input' folder.
 This is associated with test named 'test' (Test B).
}
