% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_con.R
\name{calibrate_con}
\alias{calibrate_con}
\title{calibrate_con}
\usage{
calibrate_con(test, grdIntVec, cdLst, pid, grdNm, covNmVec)
}
\arguments{
\item{test}{Name of test.}

\item{grdIntVec}{Integer vector of year levels. This is used with argument 'test'
to read data in 'data' folder.}

\item{cdLst}{List of dataframe (Item, Key, Max_score) for each year level.}

\item{pid}{Name of candidates' ID variable.}

\item{grdNm}{Variable name of Grade in data.}

\item{covNmVec}{Name vector of covariates before responses.}
}
\description{
This function merges grade datasets and codebook list, regress on grade, and
performs a free concurrent calibration. This is associated with test named 'test'.
}
