% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_concurrent.R
\name{calibrate_concurrent}
\alias{calibrate_concurrent}
\title{calibrate_concurrent}
\usage{
calibrate_concurrent(test, grades, cdLst, pid, grade, covriables)
}
\arguments{
\item{test}{Name of test.}

\item{grades}{Integer vector of year levels. This is used with argument 'test'
to read data in 'data' folder.}

\item{cdLst}{List of dataframe (Item, Key, Max_score) for each year level.}

\item{pid}{Name of candidates' ID variable.}

\item{grade}{Variable name of Grade in data.}

\item{covriables}{Name vector of covariates before responses.}
}
\description{
This function merges grade datasets and codebook list, regress on grade, and
performs a free concurrent calibration. This is associated with test named 'test'.
}
