% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_lst.R
\name{calibrate_lst}
\alias{calibrate_lst}
\title{calibrate_lst}
\usage{
calibrate_lst(dataLst = NULL, keysLst, pid, n_cov, run)
}
\arguments{
\item{dataLst}{List of response dataframe. Default is NULL (reruns after review).
If NULL and items need collapsing, put processed data with test name in
'data' folder before running.}

\item{keysLst}{List of key dataframes. Element names are test names.}

\item{pid}{Name of candidates' ID variable.}

\item{n_cov}{Number of covariates before responses.}

\item{run}{Integer that indicates round of run.}
}
\description{
This function calibrates a list of tests and put results into a summary file
in 'results' folder with 'itn' and #run in name.
}
