% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCC_fine_flag4.R
\name{CCC_fine_flag4}
\alias{CCC_fine_flag4}
\title{CCC_fine_flag4}
\usage{
CCC_fine_flag4(iflag123, comments)
}
\arguments{
\item{iflag123}{Item statistics with flag priorities of 1 to 3.}

\item{comments}{Dataframe with Comments.}
}
\value{
Item statistics with four levels of priority.
}
\description{
This function adds comments by comparing flags of priorities 1 to 3 and comments only based on CCC. If item statistics show issues but CCC are fine, comments indicate so. Also, if CCC shows issues but item statistics are fine, comments indicated so.
}
