% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCC_ipMap.R
\name{CCC_ipMap}
\alias{CCC_ipMap}
\title{CCC_ipMap}
\usage{
CCC_ipMap(
  run,
  test,
  cqs,
  abilEst2use = "pv1",
  numAbilGrps = NULL,
  poly_key = FALSE,
  anchor = FALSE,
  savePlots = TRUE,
  plotItemi = NULL
)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of the test.}

\item{cqs}{CQS output file from ConQuest.}

\item{abilEst2use}{Ability type used for curve data. Default is 'pv1'.
Use 'wle' for smaller samples.}

\item{numAbilGrps}{Number of ability groups. Default is NULL.}

\item{poly_key}{TRUE if the key of any item has polytomous scoring. Default is FALSE.}

\item{anchor}{TRUE if anchoring is being done. Default is FALSE.}

\item{savePlots}{TRUE to save both plots to calibration/'run' folder. Default is TRUE}

\item{plotItemi}{Integer to indicate item order in test. NULL to plot all items.}
}
\value{
Plots of CCC by category and score.
}
\description{
This function draws for each item CCC of both category and score.
Plots will be saved in 'output' folder.
}
\examples{
plot_data <- CCC_ipMap(test, cqs)
plot_data <- CCC_ipMap(test='DIP', cqs=cqs, poly_key=TRUE)
}
