% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCC_plot.R
\name{CCC_plot}
\alias{CCC_plot}
\title{CCC_plot}
\usage{
CCC_plot(easy = 90, hard = 10, iRst = 0.11, fit_w = 1.1, fit_uw = 1.2)
}
\arguments{
\item{easy}{Threshold to flag easy items. Default is 90 (percent correct).}

\item{hard}{Threshold to flag hard items. Default is 10 (percent correct).}

\item{iRst}{Threshold to flag low item-rest correlation statistics.
Default is 0.11.}

\item{fit_w}{Threshold to flag large weighted item fit statistics.
Default is 1.1.}

\item{fit_uw}{Threshold to flag large unweighted item fit statistics.
Default is 1.2.}
}
\description{
This function returns a list of functions that plot CCC for dichotomous,
polytomous, and right-wrong items.
}
