% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCC_Vernon.R
\name{CCC_Vernon}
\alias{CCC_Vernon}
\title{CCC_Vernon}
\usage{
CCC_Vernon(
  folder = here::here("output"),
  test,
  abilEst2use = "pv1",
  numAbilGrps = NULL,
  long_label = FALSE
)
}
\arguments{
\item{folder}{Working directory where .thr and .cqs are located. Default is 'output' folder.}

\item{test}{Name of the test.}

\item{abilEst2use}{Ability type used for curve data. Default is 'pv1'. Use 'wle' for smaller samples.}

\item{numAbilGrps}{Number of ability groups. Default is NULL.}

\item{long_label}{Whether item labels are longer than 16 characters' fixed width. Default is FALSE.}
}
\value{
Plots of CCC by category and score.
}
\description{
This function draws for each item CCC of both category and score. Plots will be saved in 'output' folder in the working directory.
}
\examples{
plot_data <- CCC_Vernon(test='elana_poly_score', abilEst2use='pv1')
plot_data <- CCC_Vernon(test='racp1', abilEst2use='pv1')
}
