% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{check_convergence}
\usage{
check_convergence(cqs, test)
}
\arguments{
\item{cqs}{CQS output file from ConQuest.}

\item{test}{Test name.}
}
\description{
This function checks convergence of the model by checking model deviance
change and maximum changes in estimates of delta's, covariance, and regressors.
This is associated with test named 'test'. A plot of the checks is saved in
'output' folder in the working directory.
}
\examples{
check_convergence(test='AHU', cqs=cqs)
}
