% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_freq_resps_cat.R
\name{check_freq_resps_cat}
\alias{check_freq_resps_cat}
\title{check_freq_resps_cat}
\usage{
check_freq_resps_cat(run, test, resp)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of the test.}

\item{resp}{Responses to items in the test.}
}
\value{
Dataframe with information such as counts in xxx_opt.txt file, counts
from R function, and their difference.
}
\description{
This function compares frequencies of item categories from CQ .txt output
file against those calculated with Function 'freq_resps_cat'. This is
associated with test named 'test'. An Excel file with comparison results
will be save in the 'output' 'output'.
}
\examples{
check_freq_resps_cat()
}
