% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_test_step.R
\name{chi_square_test_step}
\alias{chi_square_test_step}
\title{chi_square_test_step}
\usage{
chi_square_test_step(df, desig_effect = 1)
}
\arguments{
\item{df}{Datarame of step anchors' delta and error estimates in two tests 
(.x, .y), with variable names of 'delta.x', 'delta.y', 'error.x', and 'error.y'.}

\item{desig_effect}{Value to adjust errors. Default is 1.}
}
\value{
Dataframe of chi-square test results on step estimates.
}
\description{
This function does chi-square test on two groups of delta estimates of 
item step parameters. This is used for equating between step anchors of two tests.
}
\examples{
chi_square_test_step()
}
