% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqT.R
\name{chisqT}
\alias{chisqT}
\title{chisqT}
\usage{
chisqT(df, sigma = FALSE)
}
\arguments{
\item{df}{Datarame of anchors' delta and error estimates in two tests
(.x, .y), with variable names of 'delta.x', 'delta.y', 'error.x', and 'error.y'.}

\item{sigma}{Indicator of how 'delta.y' is scaled. If TRUE, it is scaled
to have same mean and sd as 'delta.x'. If FALSE, it has same mean as 'delta.x'.
Default is FALSE.}
}
\value{
Dataframe of chi-square test results.
}
\description{
This function does chi-square test on two groups of delta estimates.
This is used for equating between anchors of two tests.
}
\examples{
chisqT(df, T)
}
