% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convg_Vernon.R
\name{convg_Vernon}
\alias{convg_Vernon}
\title{convg_Vernon}
\usage{
convg_Vernon(folder = here::here("output"), test)
}
\arguments{
\item{folder}{Where the ConQuest output .cqs file is located. Default is 'output' folder in working directory.}

\item{test}{Name of test.}
}
\description{
This function checks convergence of the model by checking model deviance change and maximum changes in estimates of delta's, covariance, and regressors. This is associated with test named 'test'. A plot of the checks is saved in 'output' folder in the working directory.
}
\examples{
convg_Vernon(test='racp1')
}
