% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_plot_File.R
\name{cqc_plot_File}
\alias{cqc_plot_File}
\title{cqc_plot_File}
\usage{
cqc_plot_File(folder = "output", test, runs)
}
\arguments{
\item{folder}{Folder where ConQuest output files are located.}

\item{test}{Name of test.}

\item{runs}{Vector of test forms, e.g., c('2A', '2B', '3A', '3B').}
}
\description{
This function creates 'test_CCC_plots.cqc' file in 'input' folder.
This output file needs to be imported into and run in the GUI version of ConQuest.
}
\examples{
cqc_plot_File()
}
