% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cqc.R
\name{create_cqc}
\alias{create_cqc}
\title{create_cqc}
\usage{
create_cqc(
  wd = here::here(),
  test,
  run,
  resps_cols,
  pid_cols,
  run_ls = NULL,
  regr_ls,
  codes,
  delete,
  anchor = FALSE,
  section_extr = NULL,
  dbl_key = NULL,
  poly_key = FALSE,
  quick = FALSE,
  step = FALSE,
  DIFVar = NULL,
  DIFVar_cols,
  poly_catgrs = NULL,
  poly_facet = FALSE,
  poly_group = FALSE
)
}
\arguments{
\item{wd}{Working directory. Default is the folder where .Rproj is located.}

\item{test}{Name of the test.}

\item{run}{Vector of specific categories of variables to select from 'test_Data.txt' in 'data' folder, e.g., c('English2', 3, 1). This corresponds to the previous argument `run_ls`.}

\item{resps_cols}{Column numbers of responses.}

\item{pid_cols}{Column numbers of person ID.}

\item{run_ls}{List of data filters. Element is column number in data. Element name is filter variable's name. This argument is useful for data including more than one test forms each of which has respective keys and labels. The elements will be used by 'Keepcases' to filter a test form.}

\item{regr_ls}{List of regressors. Element is column number in data. Element name is regressor's name.}

\item{codes}{Vector of valid codes for item responses, e.g., c(1, 2, 3, 4, 5, 6, 7, 8, 9).}

\item{delete}{Vector of item order number to be removed from the test, e.g., c(2, 3, 45, 46).}

\item{anchor}{TRUE when anchor is to be done. Default is FALSE.}

\item{section_extr}{Extra sections to be added to 'test.cqc' file in 'input' folder.}

\item{dbl_key}{TRUE if any item has polytomous scoring. Default is NULL.}

\item{poly_key}{TRUE if the key of any item has polytomous scoring. Default is FALSE.}

\item{quick}{TRUE if quick estimation is preferred. Default is FALSE.}

\item{step}{TRUE if any item in the test has polytomous scoring. Default is FALSE.}

\item{DIFVar}{Name of DIF variable. Should be lowercase for ConQuest to run.}

\item{DIFVar_cols}{DIF variable's column number in data.}

\item{poly_catgrs}{Vector of polytomous DIF variable's categories.}

\item{poly_facet}{TRUE if facet model is to be run on a polytomous DIF variable. Default is FALSE.}

\item{poly_group}{TRUE if model is run per group. Default is FALSE.}
}
\description{
This function creates 'test.cqc' file in 'input' folder. This is associated with test named 'test'.
}
\details{
For DIF analysis on dichotomous variables, you need specify 'DIFVar', 'DIFVar_cols'.
For DIF analysis on polytomous variables, there are two methods. Method 1 compares each category's delta estimates with the mean of all subgroups. Specified arguments are 'DIFVar', 'DIFVar_cols', 'DIFVar_cols'. Method 2 runs two models. The first model does calibrations on each subgroup separately where you need to specify arguments of 'DIFVar', 'DIFVar_cols', and 'poly_group'. The other model runs a facet model with an interaction term between group and item where you need to specify arguments of 'DIFVar', 'DIFVar_cols', and 'poly_facet'.
}
\examples{
create_cqc()
}
