% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{create_folders}
\usage{
create_folders(DIFVar = NULL)
}
\arguments{
\item{DIFVar}{Name of DIF variable for DIF analysis. Default is NULL. When a
name is provided, a folder with that name will be created under 'DIF' folder
to hold DIF analysis output files and analysis results.}
}
\description{
This function creates folders for a pineline of psychometric analyses.
}
\details{
This function create folders to put different types of files. The folders
are named as 'data', 'dataRaw', 'DIF', 'equating', 'input', 'output', 'rCode',
and 'results'. 'dataRaw' contains raw dataset files received. 'data' contains
processed dataset, key, and label files (.txt) that correspond to a test
such as 'HUM_1' (Humanities, Grade 1). 'DIF' and 'equating' contain table
and plot results (.xlsx, .pdf) of DIF and equating analyses. 'input' contains
ConQuest input files of control files (.cqc), label files (.lab), and perhaps
anchor files (.anc). 'output' contains output files from ConQuest. 'rCode'
contains R code files (.R). 'results' contains summary files.
}
