% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_into_Data.R
\name{data_into_Data}
\alias{data_into_Data}
\title{data_into_Data}
\usage{
data_into_Data(test = NULL, data)
}
\arguments{
\item{test}{Name of test. Default is NULL.}

\item{data}{Dataframe where responses are put together.}
}
\description{
This function saves dataset 'test_Data.txt' into 'data' folder. The saved dataset is imported into ConQuest later for modeling. This is associated with test named 'test'.
}
\examples{
data_into_Data()
}
