% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_DIF_dich_step.R
\name{delta_DIF_dich_step}
\alias{delta_DIF_dich_step}
\title{delta_DIF_dich_step}
\usage{
delta_DIF_dich_step(folder = NULL, test, DIFVar, quick = TRUE)
}
\arguments{
\item{folder}{Folder where subgroups' calibration results are located. Default is the subfolder with name of argument 'DIFVar' within 'DIF' folder (folder=NULL).}

\item{test}{Name of test.}

\item{DIFVar}{Name of dichotomous DIF variable, e.g., 'gender'.}

\item{quick}{Whether quick error is needed. Default is TRUE for DIF analysis.}
}
\value{
Dataframe of variables of 'delta.x', 'delta.y'.
}
\description{
This function extracts delta statistics for two categories of dichotomous DIF variable.
}
\examples{
delta_DIF_dich_step(test='Writing', DIFVar='gender')
}
