% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_error_DIF_dich.R
\name{delta_error_DIF_dich}
\alias{delta_error_DIF_dich}
\title{delta_error_DIF_dich}
\usage{
delta_error_DIF_dich(folder, test)
}
\arguments{
\item{folder}{Folder where xxx_shw.txt file is located with delta estimates for
interaction term between dichotomous DIF variable and item. Default is 'DIF' folder.}

\item{test}{Name of test.}
}
\value{
Dataframe of three variables of 'item', 'error.x', 'error.y'.
}
\description{
This function extracts error statistics for two categories of dichotomous
DIF variable.
}
\examples{
delta_error_DIF_dich(folder=folder, test='RandD2')
}
