% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaStepTransform.R
\name{deltaStepTransform}
\alias{deltaStepTransform}
\title{deltaStepTransform}
\usage{
deltaStepTransform(test)
}
\arguments{
\item{test}{Name of test.}
}
\value{
Dataframe of variables 'item', 'delta', 'error'.
}
\description{
This function transforms step Tau estimates to deltas by adding each item
estimate to that item's step Tau estimates. For quick method, item deltas are
first adjusted by their mean to center at 0.
}
\examples{
deltaStepTransform('WA')
}
