% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_del.R
\name{df_del}
\alias{df_del}
\title{df_del}
\usage{
df_del(run, test)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of test.}
}
\value{
Dataframe of item name_step and delta.
}
\description{
This function extracts step parameter estimates from .del file. This is
associated with test named 'test'.
}
\examples{
df_del(test='elana_math_NSW')
df_del(test='elana_math_NSW', long_label=TRUE)
}
