% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_del_shw_Step.R
\name{df_del_shw_Step}
\alias{df_del_shw_Step}
\title{df_del_shw_Step}
\usage{
df_del_shw_Step(run, test)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of test.}
}
\value{
Dataframe of item name_step, delta, and error.
}
\description{
This function extract item steps' delta estimates from .shw file. This is
associated with test named 'test'.
}
\examples{
df_del_shw_Step(test='ELNA')
}
