% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_its.R
\name{df_its}
\alias{df_its}
\title{df_its}
\usage{
df_its(folder = "output", test)
}
\arguments{
\item{folder}{Folder where ConQuest output files are located.}

\item{test}{Name of test.}
}
\value{
Dataframe of item statistics such as N, facility, and discrimination.
}
\description{
This function extracts parameter estimates from its.txt file. This is
associated with test named 'test'.
}
\examples{
df_its(test='Bang_3')
}
