% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_shw_Step.R
\name{df_shw_Step}
\alias{df_shw_Step}
\title{df_shw_Step}
\usage{
df_shw_Step(run, test)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of test.}
}
\description{
This function extracts step parameter estimates from .shw file. This is
associated with test named 'test'.
}
\examples{
df_shw_Step(test='ELNA')
}
