% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_comment_dich.R
\name{DIF_comment_dich}
\alias{DIF_comment_dich}
\title{DIF_comment_dich}
\usage{
DIF_comment_dich(DIFVar, iDIF)
}
\arguments{
\item{DIFVar}{Name of DIF variable.}

\item{iDIF}{Vector of DIF items.}
}
\description{
This function generates comments for DIF analysis on dichotomous variable. 
This is associated with test named 'test'.
}
\examples{
DIF_comment_dich()
}
