% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_comment_dich_equate.R
\name{DIF_comment_dich_equate}
\alias{DIF_comment_dich_equate}
\title{DIF_comment_dich_equate}
\usage{
DIF_comment_dich_equate(vars, iDIF, DIFVar)
}
\arguments{
\item{vars}{Vector of length 2 such as c('girls','boys'). Its order
corresponds to the alphabetic/numeric order of DIF variables' two categories in data.}

\item{iDIF}{Vector of DIF items.}

\item{DIFVar}{Name of dichotomous DIF variable, e.g., 'gender'.}
}
\description{
This function generates comments for DIF analysis on dichotomous variable.
This is associated with test named 'test'.
}
