% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_comment_poly.R
\name{DIF_comment_poly}
\alias{DIF_comment_poly}
\title{DIF_comment_poly}
\usage{
DIF_comment_poly(DiF_sum)
}
\arguments{
\item{DiF_sum}{Dataframe of summary of results from a Bonferroni-adjusted statistical test.}
}
\value{
Tibble of summary of results from polytomous DIF variable analysis.
}
\description{
This function summarizes results from a type of DIF analysis on a polytomous DIF variable.
}
\examples{
DIF_comment_poly()
}
