% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_dich.R
\name{DIF_dich}
\alias{DIF_dich}
\title{DIF_dich}
\usage{
DIF_dich(
  folder = NULL,
  DIFVar,
  test,
  vars_dich,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_adj_cut = 5,
  chi_cut = 10,
  facil_cut = 10,
  steps = FALSE,
  sav_xlsx = TRUE
)
}
\arguments{
\item{folder}{Folder where subgroups' calibration results are located. Default is the subfolder with name of argument 'DIFVar' within 'DIF' folder (folder=NULL).}

\item{DIFVar}{Name of dichotomous DIF variable, e.g., 'gender'.}

\item{test}{Name of test.}

\item{vars_dich}{Vector of categories of DIFVar, e.g., c('girl', 'boy'). This is used in plotting. The order should be the same as '1' and '2' in interaction term of ConQuest's .shw files. Note that ConQuest follows alphabetical rule and uses '1' and '2' in output files for 'F' and 'M' (coded in data for female and male).}

\item{p_cut}{P value for performing chi-square significance tests. One criterion to flag DIF items. Default is 0.05.}

\item{DIF_cut}{Threshold of difference between two categories' delta estimates. One criterion to flag DIF items. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of adjusted difference between two categories' delta estimates. One criterion to flag DIF items. Default is 5.}

\item{chi_cut}{Chi-square value from chi-square test. One criterion to flag DIF items. Default is 10.}

\item{facil_cut}{Threshold of number of percent to flag an item with large facility difference between two groups of test takers. Default is 10.}

\item{steps}{TRUE if the chi-square test is on step parameters of polytomous items. Default is FALSE.}

\item{sav_xlsx}{TRUE if an Excel file of DIF analysis results are to be saved. Default is FALSE.}
}
\value{
Dataframe of summary of results from dichotomous DIF variable analysis.
}
\description{
This function performs DIF analyses on a dichotomous variable and summarizes results.
}
\examples{
DIF_dich()
}
