% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_dich_iterative.R
\name{DIF_dich_iterative}
\alias{DIF_dich_iterative}
\title{DIF_dich_iterative}
\usage{
DIF_dich_iterative(
  DIFVar,
  test,
  vars,
  df,
  p_cut = 0.05,
  chi_cut = 10,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  desig_effect = 1,
  steps = FALSE,
  facil_cut = 10,
  plot_facil = TRUE,
  long_label = FALSE
)
}
\arguments{
\item{DIFVar}{Name of dichotomous DIF variable, e.g., 'gender'.}

\item{test}{Name of test.}

\item{vars}{Vector of length 2 such as c('girls','boys'). Its order corresponds to the alphabetic/numeric order of DIF variables' two categories in data.}

\item{df}{Dataframe with delta estimates and errors of items for both categories of dichomomous variable in a test.}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{chi_cut}{Threshold of chi-square difference between two tests. Default is 10.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference between two tests. Default is 4.}

\item{desig_effect}{Value to adjust errors. Default is 1.}

\item{steps}{TRUE if DIF analysis is performed on step parameters. Default is FALSE.}

\item{facil_cut}{Threshold of number of percent to flag an item with large facility difference between two groups of test takers. Default is 10.}

\item{plot_facil}{TRUE if a facility plot is desired of delta estimates for the two DIF variable categories.}

\item{long_label}{Whether item labels are longer than 11 characters' fixed width. Default is FALSE.}
}
\value{
List of summary of results from dichotomous DIF variable analysis, including comments, steps, summary statistics with flags, and statistics of items after review.
}
\description{
This function performs chi-square tests (DIF analysis) on all items's difference of delta estimates between two groups of test takers. An Excel file with test results and flags is saved in 'DIF' folder. Also, one plot is saved in subfolder 'plot' inside 'DIF' folder.
}
\examples{
DIF_dich_iterative(DIFVar='gender', test='N_1', vars=c('Girls', 'Boys'))
}
