% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_dim_one.R
\name{DIF_dim_one}
\alias{DIF_dim_one}
\title{DIF_dim_one}
\usage{
DIF_dim_one(
  method = c("chi_square", "Bonferroni", "Facet"),
  test,
  pid,
  n_cov,
  DIFVar,
  data = NULL,
  keys = NULL,
  wd = here::here(),
  vars = NULL,
  poly_facet = FALSE,
  n_resp = NULL,
  regr_vec_char = NULL,
  labels = NULL,
  quick = FALSE,
  section_extr = NULL,
  miss_code = c(".", "r", "R", "x", "X", "", " ")
)
}
\arguments{
\item{method}{One of 'chi_square', 'Bonferroni', or 'Facet'.}

\item{test}{Name of the test.}

\item{pid}{Name of candidates' ID variable.}

\item{n_cov}{Number of covariates before responses.}

\item{DIFVar}{Name of DIF variable. Should be lowercase for ConQuest to run. Default is NULL.}

\item{data}{Dataframe with pid, covariables (e.g,, DIF variable), and responses. Default is NULL where Excel file with name 'test' in 'data' folder is used.}

\item{keys}{Vector of keys in the test. Default is NULL.}

\item{wd}{Working directory. Default is the folder where .Rproj is located.}

\item{vars}{Vector of length 2 such as c('girls','boys'). Its order corresponds to the alphabetic/numeric order of DIF variables' two categories in data.}

\item{poly_facet}{TRUE if facet model is to be run on a polytomous DIF variable. Default is FALSE.}

\item{n_resp}{Number of items. Default is NULL.}

\item{regr_vec_char}{Vector of character regressors' names.}

\item{labels}{Vector of item labels that correspond to order of item response columns in data.}

\item{quick}{TRUE when testing. Default is FALSE.}

\item{section_extr}{Extra sections to be added to .cqc file in 'input' folder. Default is NULL.}

\item{miss_code}{Code for missing data that should be converted to NA. Default is c('.', 'r', 'R', 'x', 'X', '', ' ').}
}
\description{
This function can perform DIF analysis on either dichotomous or polytomous variable. This is associated with test named 'test'.
}
\details{
Should preprocess data in three ways. First, convert non-used categories and missing values to NA. Second, remove item responses with all missing values. Third, remove items that have all missing values or all correct/incorrect responses on any category of the DIF variable.
}
\examples{
DIF_dim_one(method='chi_square', vars=c('Girls', 'Boys'), data=df_DIF, test='David', DIFVar="studGender", pid="studentId", n_cov=5, regr_vec_char=c('countryName', 'studAge'), quick=TRUE, miss_code='.')
DIF_dim_one(method='Bonferroni', data=df_DIF[, 1:100], test='David', DIFVar="studAge", pid="studentId", n_cov=5, quick=TRUE, miss_code='.')
DIF_dim_one(method='Facet', data=df_DIF[, 1:100], test='David', DIFVar="studAge", pid="studentId", n_cov=5, quick=TRUE, miss_code='.')
DIF_dim_one(method='chi_square', vars=c('Girls', 'Boys'), data=elena, test='elena1', DIFVar="gender", pid="IDSTUD", n_cov=9, regr_vec_char=c('province', 'quintile'), quick=TRUE, miss_code=NULL)
DIF_dim_one(method='Bonferroni', data=elena, test='Elena1', DIFVar="quintile", pid="IDSTUD", n_cov=9, regr_vec_char=c('province', 'gender'), quick=TRUE, miss_code=NULL)
DIF_dim_one(method='Facet', data=elena, test='Elena1', DIFVar="quintile", pid="IDSTUD", n_cov=9, regr_vec_char=c('province', 'gender'), quick=TRUE, miss_code=NULL)
}
