% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_domain.R
\name{DIF_domain}
\alias{DIF_domain}
\title{DIF_domain}
\usage{
DIF_domain(
  test,
  data = NULL,
  var_ls = NULL,
  n_cov = NULL,
  pid = NULL,
  keys = NULL,
  labels = NULL,
  poly_key = FALSE,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  facil_cut = 10,
  desig_effect = 1,
  step = FALSE,
  output_ready = FALSE,
  miss_code = c("M", "R"),
  iterative = TRUE,
  quick = TRUE
)
}
\arguments{
\item{test}{Name of the test.}

\item{data}{Dataframe of DIF variables and responses.}

\item{var_ls}{List of DIF variables. Name is variable name. Element is name
of categories which corresponds to data coding order.}

\item{n_cov}{Number of covariates before responses.}

\item{pid}{Name of candidates' ID variable.}

\item{keys}{Vector of keys in the test.}

\item{labels}{Vector of item labels that correspond to order of item response
columns in data.}

\item{poly_key}{TRUE if the key of any item has polytomous scoring.
Default is FALSE.}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two
tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference
between two tests. Default is 4.}

\item{facil_cut}{Threshold of number of percent to flag an item with large
facility difference between two groups of test takers. Default is 10.}

\item{desig_effect}{Value to adjust errors. Default is 1.}

\item{step}{TRUE if DIF analysis is performed on step parameters. Default
is FALSE.}

\item{output_ready}{TRUE if results are ready. Default is FALSE.}

\item{miss_code}{Missing codes. Default is c('M', 'R').}

\item{iterative}{TRUE to iteratively remove DIF items. Default is TRUE.}

\item{quick}{Whether quick error is needed. Default is TRUE for DIF analysis.}
}
\description{
This function performs DIF analysis on variables and responses in the data.
Data should be processed to have DIF variables of only categories needed.
This is associated with test named 'test'. An Excel file with a summary of
the DIF analysis results and plots will be saved in 'DIF' folder in the
working directory.
}
\examples{
DIF_domain()
}
