% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_poly_shw.R
\name{DIF_poly_shw}
\alias{DIF_poly_shw}
\title{DIF_poly_shw}
\usage{
DIF_poly_shw(
  folder = NULL,
  DIFVar,
  labels,
  test,
  domain = NULL,
  p_cut = 0.05,
  steps = FALSE
)
}
\arguments{
\item{folder}{Folder where subgroups' calibration results are located. Default is the subfolder with name of argument 'DIFVar' within 'DIF' folder (folder=NULL). It's advisable to put factor level in file names such as 'test_4.shw' and put all groups' .shw files in this folder.}

\item{DIFVar}{Name of polytomous DIF variable, e.g., 'nation'.}

\item{labels}{Dataframe of labels of all the items in a test. Variable 'item' is item order in label file ('test_Labels.txt' in 'data' folder) and in test. Variable 'label' is complete item labels of test items.}

\item{test}{Name of the test.}

\item{domain}{Name of the domain in the test, e.g., 'Literacy'. Default is NULL.}

\item{p_cut}{Threshold of statistical test. Default is 0.05.}

\item{steps}{TRUE if polytomous items are involved. Default is FALSE.}
}
\value{
List of summary of results from polytomous DIF variable analysis, including comments, steps, summary statistics with flags, and statistics used to produce the the flags.

Tibble of summary of results from polytomous DIF variable analysis.
}
\description{
This function extracts item statistics from subgroups of the polytmous DIF variable (in 'test_category.shw' files in DIF variable's subfolder within 'DIF' folder) to produce a list of dataframes of those statistics. Then it calls function DIF_poly() to perform a Bonferroni-adjusted statistical test and flag the combinations of item and category that show significant results.
}
\examples{
DIF_poly_shw(DIFVar='quintile', labels=labels, test=test)
}
