% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_steps_dich.R
\name{DIF_steps_dich}
\alias{DIF_steps_dich}
\title{DIF_steps_dich}
\usage{
DIF_steps_dich(iterative = FALSE)
}
\arguments{
\item{iterative}{TRUE to iteratively remove DIF items. Default is FALSE}
}
\value{
Tibble of six steps to do DIF analyses on dichotomous variable.
}
\description{
This function generates a tibble of steps to follow when performing DIF 
analysis on a dichotomous DIF variable.
}
\examples{
# Not run
# DIF_steps_dich()
}
