% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_steps_dich_step.R
\name{DIF_steps_dich_step}
\alias{DIF_steps_dich_step}
\title{DIF_steps_dich_step}
\usage{
DIF_steps_dich_step(iterative = FALSE)
}
\arguments{
\item{iterative}{TRUE to iteratively remove DIF items. Default is FALSE}
}
\value{
Tibble of six steps to do DIF analyses on dichotomous variable and 
polytomous responses.
}
\description{
This function generates a tibble of steps to follow when performing DIF 
analysis on a dichotomous DIF variable and polytomous responses.
}
\examples{
# Not run
# DIF_steps_dich_step(iterative=TRUE)
}
