% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_steps_poly.R
\name{DIF_steps_poly}
\alias{DIF_steps_poly}
\title{DIF_steps_poly}
\usage{
DIF_steps_poly()
}
\value{
Tibble of four steps to do DIF analyses.
}
\description{
This function generates a tibble of steps to follow when performing a type
of DIF analysis on a polytomous DIF variable.
}
\examples{
DIF_steps_poly()
}
