% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_symbol_poly.R
\name{DIF_symbol_poly}
\alias{DIF_symbol_poly}
\title{DIF_symbol_poly}
\usage{
DIF_symbol_poly(cutscore, t_vec, cat_vec)
}
\arguments{
\item{cutscore}{Bonferroni adjustment of statistical significance value 
p (e.g., 0.05) of a two-sided test. The adjusted p value is p/(n*(n-1)/2)/2 
where n is number of categories of the polytomous DIF variable.}

\item{t_vec}{Vector of standardized difference between a subgroup's items' 
delta estimates and the average of all subgroups' items' delta estimates. 
'standardized' means that the delta difference of each item in each subgroup 
is divided by the square of corresponding error of that delta.}

\item{cat_vec}{Vector of difference between a subgroup's items' delta 
estimates and the average of all subgroups' items' delta estimates.}
}
\value{
Tibble of Bonferroni adjusted statistical significance results of 
three variables/columns. Column 1 is that subgroup's item delta difference, 
symbol of significance ('*') or not, symbol of bigger than delta mean ('B') 
or smaller ('S').
}
\description{
This function is called by function DIF_poly_df() to perform Bonferroni 
significance test of difference between each subgroup's delta and the 
average delta of all subgroups.
}
\examples{
DIF_symbol_poly()
}
