% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_VarTests.R
\name{DIF_VarTests}
\alias{DIF_VarTests}
\title{DIF_VarTests}
\usage{
DIF_VarTests(
  tests = NULL,
  dfLst = NULL,
  difVarsLst = NULL,
  n_cov = NULL,
  pid = NULL,
  keysLst = NULL,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  desig_effect = 1,
  step = FALSE,
  output_ready = FALSE,
  iterative = TRUE,
  test3term = NULL
)
}
\arguments{
\item{tests}{Vector of test names. Default is NULL.}

\item{dfLst}{Dataframe of DIF variables and responses.}

\item{difVarsLst}{List of DIF variables. Name is variable name. Element is name
of categories which corresponds to data coding order. Polymotous DIF variable
should have element as NULL. Default is NULL.}

\item{n_cov}{Number of covariates before responses. Default is NULL.}

\item{pid}{Name of candidates' ID variable. Default is NULL.}

\item{keysLst}{List of dataframe of 'Item', 'Key', and 'Max_score'
(add Key2 if double key). Default is NULL.}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two
tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference
between two tests. Default is 4.}

\item{desig_effect}{Value to adjust errors. Default is 1.}

\item{step}{TRUE if DIF analysis is performed on step parameters. Default
is FALSE.}

\item{output_ready}{TRUE if results are ready. Default is FALSE.}

\item{iterative}{TRUE to iteratively remove DIF items. Default is TRUE.}

\item{test3term}{Vector of names of tests to include a three-way interaction
(item, step, DIF variable). Default is NULL.}

\item{miss_code}{Missing codes. Default is c('M', 'R').}

\item{byDIFVar}{Whether to summarise DIF results by DIF variable. Default
is TRUE. Set byDIFVar as FALSE to summarise by test/domain.}
}
\description{
This function performs DIF analysis on variables and responses in the data.
Data should be processed to have DIF variables of only categories needed.
This is associated with test named 'test'. An Excel file with a summary of
the DIF analysis results and plots will be saved in 'DIF' folder in the
working directory.
}
\examples{
DIF_VarTests()
}
