% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIFDimMulti.R
\name{DIFDimMulti}
\alias{DIFDimMulti}
\title{DIFDimMulti}
\usage{
DIFDimMulti(
  method = c("chi_square", "Bonferroni"),
  test,
  DIFVar,
  pid,
  n_cov,
  nDimVec,
  dimNmVec,
  respDf = NULL,
  filetype = "sav",
  keyDf,
  vars = NULL,
  regrNmVec = NULL,
  quick = TRUE,
  delVec = NULL,
  dblKeyLst = FALSE,
  section_extr = NULL,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  design_effect = 1,
  dim_multi = FALSE,
  scores = NULL,
  prep_process = FALSE,
  pweight = NULL,
  iter = TRUE
)
}
\arguments{
\item{method}{One of 'chi_square' or 'Bonferroni'.}

\item{test}{Name of the test.}

\item{DIFVar}{Name of DIF variable. Should be lowercase for ConQuest to run.
Default is NULL.}

\item{pid}{Name of candidates' ID variable.}

\item{n_cov}{Number of covariates before responses.}

\item{nDimVec}{Vector of numbers of responses the dimensions have.}

\item{dimNmVec}{Vector of the dimensions' names.}

\item{respDf}{Dataframe with pid, covariables (e.g,, DIF variable), and responses.
Default is NULL where Excel file with name 'test' in 'data' folder is used.}

\item{filetype}{Format for input dataset. Default is 'sav'. Also support
csv format.}

\item{keyDf}{Dataframe of 'Item', 'Key', and 'Max_score' (add Key2 if double key).}

\item{vars}{Vector of length 2 such as c('girls','boys'). Its order
corresponds to the alphabetic/numeric order of DIF variables' two categories in data.}

\item{regrNmVec}{Vector of character regressors' names. Default is NULL.
Make sure to check colinearity of the regressors.}

\item{quick}{TRUE when testing. Default is TRUE}

\item{delVec}{Vector of orders of items to be removed.}

\item{dblKeyLst}{TRUE if any item has polytomous scoring. Default is FALSE.}

\item{section_extr}{Extra sections to be added to 'test.cqc' file in 'input'
folder. Default is NULL.}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between
two tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference
between two tests. Default is 4.}

\item{design_effect}{Value to adjust errors. Default is 1.}

\item{dim_multi}{TRUE if the model is multidimensional. Default is FALSE.}

\item{scores}{Scores possible in the test, e.g., 0:3. Default is NULL.}

\item{prep_process}{TRUE if it is needed to remove items without data on
some categories of DIF variable. Default is FALSE.}

\item{pweight}{Variable name of person weights in response dataframe. Should
be specified if weight is used for modeling. Default is NULL.}

\item{iter}{TRUE to iteratively remove DIF items. Default is FALSE.}
}
\description{
This function can perform DIF analysis on either dichotomous or polytomous
variable. This is associated with test named 'test'. When dimension is above
two, you would be continuously shown "Yes/No" nonstop. In that case, please
stop the program and run the generated .cqc file in 'input' folder with
ConQuest GUI or Console version on your computer.
}
\details{
Sparse data will be processed in three ways before being used for DIF analysis.
First, convert non-used categories and missing values to NA. Second, remove
item responses with all missing values. Third, remove items that have no data
on any category of the DIF variable. An Excel with processed data and
removed items is saved in 'data' folder. Its name has both test name and
DIF variable.
}
