% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIFDimOne.R
\name{DIFDimOne}
\alias{DIFDimOne}
\title{DIFDimOne}
\usage{
DIFDimOne(
  method = c("chi_square", "Bonferroni", "Facet"),
  test,
  pid,
  n_cov,
  DIFVar,
  respDf = NULL,
  filetype = "sav",
  keyDf,
  vars = NULL,
  poly_facet = FALSE,
  n_resp = NULL,
  regrNmVec = NULL,
  quick = TRUE,
  section_extr = NULL,
  prep_process = FALSE,
  save_xlsx = TRUE,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_std_cut = 4,
  iter = FALSE,
  step = FALSE,
  design_effect = 1,
  pweight = NULL
)
}
\arguments{
\item{method}{One of 'chi_square', 'Bonferroni', or 'Facet'.}

\item{test}{Name of the test.}

\item{pid}{Name of candidates' ID variable.}

\item{n_cov}{Number of covariates before responses.}

\item{DIFVar}{Name of DIF variable. Should be lowercase for ConQuest to run.
Default is NULL.}

\item{respDf}{Dataframe with pid, covariables (e.g,, DIF variable), and responses.
Default is NULL where Excel file with name 'test' in 'data' folder is used.}

\item{filetype}{Format for input dataset. Default is 'sav'. Also support
csv format.}

\item{keyDf}{Dataframe of 'Item', 'Key', and 'Max_score' (add Key2 if double key).}

\item{vars}{Vector of length 2 such as c('girls','boys'). Its order corresponds
to the alphabetic/numeric order of DIF variables' two categories in data.}

\item{poly_facet}{TRUE if facet model is to be run on a polytomous DIF variable.
Default is FALSE.}

\item{n_resp}{Number of items. Default is NULL.}

\item{regrNmVec}{Vector of character regressors' names.}

\item{quick}{Whether quick error is needed. Default is TRUE for DIF analysis.}

\item{section_extr}{Extra sections to be added to .cqc file in 'input' folder.
Default is NULL.}

\item{prep_process}{TRUE if it is needed to remove items without data on
some categories of DIF variable. Default is FALSE.}

\item{save_xlsx}{Whether to save summary file and plots. Default is TRUE
(one DIF variable).}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between
two tests. Default is 0.5.}

\item{DIF_std_cut}{Threshold of an item's standardized delta estimate difference
between two tests. Default is 4.}

\item{iter}{TRUE to iteratively remove DIF items. Default is FALSE}

\item{step}{TRUE if polytomous items are involved. Default is FALSE.}

\item{design_effect}{Value to adjust errors. Default is 1.}

\item{pweight}{Variable name of person weights in response dataframe. Should
be specified if weight is used for modeling. Default is NULL.}
}
\description{
This function can perform DIF analysis on either dichotomous or polytomous
variable. This is associated with test named 'test'.
}
\details{
Should preprocess data in three ways. First, convert non-used categories and
missing values to NA. Second, remove item responses with all missing values.
Third, remove items that have all missing values or all correct/incorrect
responses on any category of the DIF variable.
}
