% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equate.R
\name{Equate}
\alias{Equate}
\title{Equate}
\usage{
Equate(
  deltaDf,
  test,
  vars,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_std_cut = 4,
  sav_results = TRUE,
  design_effect = 1,
  step = FALSE,
  DIFVar = NULL,
  iter = FALSE,
  indDf = NULL,
  pointer = FALSE,
  report = FALSE
)
}
\arguments{
\item{deltaDf}{Dataframe of five variables of 'item' (anchor labels), 'delta.x',
'error.x', 'delta.y', and 'error.y'. '.x' and '.y' should correspond to order
of test names in 'vars'.}

\item{test}{Name of test such as 'AGAT'.}

\item{vars}{Vector of length 2 such as c('VIC','NSW'). Its order corresponds
to two tests associated with .x and .y.}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two
tests. Default is 0.5.}

\item{DIF_std_cut}{Threshold of an item's standardized delta estimate difference
between two tests. Default is 4.}

\item{sav_results}{TRUE if an Excel file with chi-square test results and
a plot are desired. Default is TRUE.}

\item{design_effect}{Value to adjust errors. Default is 1.}

\item{step}{TRUE if DIF analysis is performed on step parameters.
Default is FALSE.}

\item{DIFVar}{Name of dichotomous DIF variable, e.g., 'gender'.}

\item{iter}{TRUE to iteratively remove DIF items. Default is FALSE.}

\item{indDf}{Dataframe of indice which includes variables of 'item' and
1-3 of facility, discrimination, and weighted fit in order. Default is NULL when
none of those indice is available.}

\item{pointer}{Whether to print locations of output files. Default is FALSE.}

\item{report}{Whether to generate a html report. Default is FALSE.}
}
\value{
Dataframe of chi-square test results for anchors between two tests.
}
\description{
This function performs chi-square tests (DIF analysis) on a group of anchor
items' difference of delta estimates in two tests. An Excel file with test
results and flags can be saved in 'equating' folder. Also, one plot is saved
in subfolder 'plot' inside 'equating' folder.
}
\examples{
Equate(deltaDf=data[, c('item', 'delta.x', 'error.x', 'delta.y', 'error.y')],
test='Elana_math', vars=c('NSW', 'VIC'))
}
