% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equate_Hrz.R
\name{Equate_Hrz}
\alias{Equate_Hrz}
\title{Equate_Hrz}
\usage{
Equate_Hrz(
  folder = here::here("output"),
  test,
  grades = c(2:10),
  forms = c("A", "B"),
  p_cut = 0.05,
  chi_cut = 10,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  steps = FALSE,
  long_label = FALSE
)
}
\arguments{
\item{folder}{Folder where all grades' calibration results on both Form A and Form B are located. Default is 'output' folder.}

\item{test}{Name of test.}

\item{grades}{Vector of all grades. Default is c(2:10).}

\item{forms}{Forms used in a test. Default is c('A','B')}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{chi_cut}{Threshold of chi-square difference between two tests. Default is 10.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference between two tests. Default is 4.}

\item{steps}{TRUE if DIF analysis is performed on step parameters. Default is FALSE.}

\item{long_label}{Whether item labels are longer than 15 characters' fixed width. Default is FALSE.}
}
\value{
List of chi-square test results on Form A and Form B for each grade.
}
\description{
This function performs horizontal equating. An Excel file with 'Equating_Horizontal' in name is saved in 'equating' folder, each sheet of which is for one grade. Also, one plot is saved in subfolder 'plot' inside 'equating' folder for each grade.
}
\examples{
Equate_Hrz(test='ArabicA', grades=c(3:6))
}
