% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_lst.R
\name{equate_lst}
\alias{equate_lst}
\title{equate_lst}
\usage{
equate_lst(
  statsLst,
  vars,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  desig_effect = 1,
  step = FALSE,
  iterative = TRUE
)
}
\arguments{
\item{statsLst}{List of tests' anchor statistics dataframe of variables 'item',
'delta.x', 'error.x', 'delta.y', 'error.y'. The names of the elements are
test names.}

\item{vars}{Vector of length 2 such as c('2022','2023'). Its order corresponds
to two tests associated with .x and .y.}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between
two tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference
between two tests. Default is 4.}

\item{desig_effect}{Value to adjust errors. Default is 1.}

\item{step}{TRUE if DIF analysis is performed on step parameters.
Default is FALSE.}

\item{iterative}{TRUE to iteratively remove DIF items. Default is TRUE.}
}
\value{
List of chi-square test results on Form A and Form B for each grade.
}
\description{
This function returns a list of functions to add format such as header and
body style to Excel sheet.
}
