% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equate_Vrt.R
\name{Equate_Vrt}
\alias{Equate_Vrt}
\title{Equate_Vrt}
\usage{
Equate_Vrt(
  test,
  grades = c(2:10),
  grade_name = "L",
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  step = FALSE,
  iterative = TRUE
)
}
\arguments{
\item{test}{Name of test.}

\item{grades}{Vector of all grades. Default is c(2:10).}

\item{grade_name}{Character to add before grade in plots. Default is 'L'
(for 'level').}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two
tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference
between two tests. Default is 4.}

\item{step}{TRUE if DIF analysis is performed on step parameters.
Default is FALSE.}

\item{iterative}{TRUE to iteratively remove DIF items. Default is TRUE}
}
\value{
List of chi-square test results for anchors between two adjacent grades.
}
\description{
This function performs vertical equating. An Excel file named 'Vrt_xxx.xlsx'
is saved in 'equating' folder, each sheet of which is for two
adjacent grades. Also saved are scatterplots named 'Vrt_xxx.pdf'.
}
\examples{
Equate_Vrt(test='Math', grades=1:3)
}
