% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equateLst.R
\name{equateLst}
\alias{equateLst}
\title{equateLst}
\usage{
equateLst(
  deltaDfLst,
  vars,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_std_cut = 4,
  design_effect = 1,
  step = FALSE,
  iter = TRUE,
  indDfLst = NULL,
  sigmaLgt = FALSE
)
}
\arguments{
\item{deltaDfLst}{List of tests' anchor statistics dataframe of variables 'item',
'delta.x', 'error.x', 'delta.y', 'error.y'. The names of the elements are
test names.}

\item{vars}{Vector of length 2 such as c('2022','2023'). Its order corresponds
to two tests associated with .x and .y.}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between
two tests. Default is 0.5. Give vector of cuts if tests vary in cut.}

\item{DIF_std_cut}{Threshold of an item's standardized delta estimate difference
between two tests. Default is 4. Give vector of cuts if tests vary in cut.}

\item{design_effect}{Value to adjust errors. Default is 1.}

\item{step}{TRUE if DIF analysis is performed on step parameters.
Default is FALSE.}

\item{iter}{TRUE to iteratively remove DIF items. Default is TRUE.}

\item{indDfLst}{List of tests' indice statistics dataframe of variables 'item',
'facil.x', 'discr.x', 'fitw.x', 'facil.y', 'discr.y', 'fitw.y'. Those indice
variable should be in the order above and can be incomplete. The names of
the elements are test names and should be in same order as deltaDfLst. Default
is NULL when no index is available.}

\item{sigmaLgt}{Indicator vector of how 'delta.y' is scaled. If TRUE, it is scaled
to have same mean and sd as 'delta.x'. If FALSE, it has same mean as 'delta.x'.
Give vector of TRUE/FALSE if tests vary in scaling method.}
}
\value{
Summary file of chi-square test results on tests in list with plots,
hyperlinks, flag color, and format.
}
\description{
This function returns a list of functions to add format such as header and
body style to Excel sheet.
}
