% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiva_tbl.R
\name{equiva_tbl}
\alias{equiva_tbl}
\title{equiva_tbl}
\usage{
equiva_tbl(
  wd = here::here(),
  test,
  est_type = "wle",
  slope = 10,
  intercept = 100,
  extrapolation = TRUE
)
}
\arguments{
\item{wd}{Working directory. Default is the folder where .Rproj is located.}

\item{test}{Name of test.}

\item{est_type}{Type of ability estimate to use, including 'wle', 'mle', 'pv1' 'pv2', 'pv3', 'pv4', and 'pv5'. Default is 'wle'.}

\item{slope}{Slope to multiply ability estimates. Default is 10.}

\item{intercept}{Value/intercept to add to ability estimates. Default is 100.}

\item{extrapolation}{Whether to extrapolate the minimum and maximum estimates. Default is TRUE.}
}
\value{
Dataframe of scaled ability estimates.
}
\description{
This function extracts desired type of ability estimates from compressed ConQuest .cqs file and generate scaled scores with specified slope and intercept. This is associated with test named 'test'.
}
\examples{
equiva_tbl()
}
