% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_band.R
\name{error_band}
\alias{error_band}
\title{error_band}
\usage{
error_band(df)
}
\arguments{
\item{df}{Dataframe with results of chi-square tests on items that 
appeared in two tests.}
}
\value{
Dataframe of chi-square test results with smoothed upper and lower 
confidence interval values added.
}
\description{
This function produces smoothed upper and lower confidence interval (CI) 
values of the average of an item's delta estimates in two tests. Those 
values will be used in plotting.
}
