% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cas.R
\name{est_cas}
\alias{est_cas}
\title{est_cas}
\usage{
est_cas(folder = here::here("output"), test)
}
\arguments{
\item{folder}{Folder where 'test.cas' file is located. Default is 'output' folder.}

\item{test}{Name of test.}
}
\value{
Dataframe with variables of 'Pid', 'ScoreRaw', 'ScoreMax', 'Est', and 'SE'.
}
\description{
This function extracts ability estimation from 'test.cas' file.
This is associated with test named 'test'.
}
\examples{
# Not run
# est_cas(test='HUM_2')
}
