% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_its.R
\name{file_its}
\alias{file_its}
\title{file_its}
\usage{
file_its(folder = "output", test, DIFVar = NULL)
}
\arguments{
\item{folder}{Folder where ConQuest output files are located.}

\item{test}{Name of test.}

\item{DIFVar}{Name of variable to perform DIF analysis on.}
}
\value{
Dataframe of delta, facility, etc..
}
\description{
This function extracts item analysis statistics from CQ output file 'test.its'.
This is associated with test named 'test'.
}
\examples{
# Not run
# file_its(test=t1)
}
