% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path_type.R
\name{file_path_type}
\alias{file_path_type}
\title{file_path_type}
\usage{
file_path_type(folder, test, type)
}
\arguments{
\item{folder}{Folder where .shw file is located with delta estimates for interaction term between dichotomous DIF variable and item.}

\item{test}{Name of test.}

\item{type}{Type of output file from ConQuest such as 'txt', 'its', 'shw', and 'itn'}
}
\value{
Path of the file.
}
\description{
This function generates file path to a specified type. This is associated with test named 'test'.
}
\examples{
file_path_type(folder=folder, test=test, type)
}
