% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_shw.R
\name{file_shw}
\alias{file_shw}
\title{file_shw}
\usage{
file_shw(folder, test)
}
\arguments{
\item{folder}{Folder where .shw file is located.}

\item{test}{Name of test.}
}
\value{
String of lines in 'test.shw' file.
}
\description{
This function reads .shw file. This is associated with test named 'test'.
}
\examples{
file_shw()
}
