% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{fit_stats}
\alias{fit_stats}
\title{fit_stats}
\usage{
fit_stats(test)
}
\arguments{
\item{test}{Name of test.}
}
\description{
This function extracts fit statistics from xxx_shw.txt file. This is
associated with test named 'test'.
}
\examples{
fit_stats(test)
}
