% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_all_scores.R
\name{freq_all_scores}
\alias{freq_all_scores}
\title{freq_all_scores}
\usage{
freq_all_scores(vec, max_lev)
}
\arguments{
\item{vec}{Vector of all scores in the test.}

\item{max_lev}{Maximum score in the test.}
}
\value{
Dataframe with raw scores, frequencies, cumulative frequencies, and cumulative percentiles.
}
\description{
This function reads in a vector of raw scores and return a dataframe of raw scores, frequencies, cumulative frequencies, and cumulative percentiles.
}
\details{
This function uses 0 as minimum score and allows users to input a maximum score of the test. Floor is used to obtain whole numbers for cumulative percentiles.
}
\examples{
freq_all_scores(vec=sample(c(0:40), 5000, replace = TRUE), max_lev=42)
}
