% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_resps_cat.R
\name{freq_resps_cat}
\alias{freq_resps_cat}
\title{freq_resps_cat}
\usage{
freq_resps_cat(
  resp,
  wide = FALSE,
  prop = FALSE,
  sort_cats = TRUE,
  na_label = "(NA)",
  digits = 2,
  freq_suffix = "_freq",
  prop_suffix = "_prop"
)
}
\arguments{
\item{resp}{Dataframe of responses where each column is a vector of responses
to one test item.}

\item{wide}{TRUE if wide form is desired. Default is FALSE. When default,
the output has long format same as frequency table in .txt file output
from ConQuest.}

\item{prop}{TRUE if proportion is desired. Default is FALSE.}

\item{sort_cats}{sort category columns in wide output.}

\item{na_label}{label to use for NA category in outputs.}

\item{digits}{rounding digits for proportions.}

\item{freq_suffix}{suffix for frequency columns in wide output.}

\item{prop_suffix}{suffix for proportion columns in wide output.}
}
\value{
Table of each item's category frequencies.
}
\description{
This function . This is associated with test named 'test'.
}
\examples{
freq_resps_cat(resp, TRUE, TRUE)
}
