% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dummy_data.R
\name{get_dummy_data}
\alias{get_dummy_data}
\title{get_dummy_data}
\usage{
get_dummy_data(num_items)
}
\arguments{
\item{num_items}{Number of items in the test.}
}
\description{
This function generate dummy data so that there is one case for one possible
total raw score in the test.
}
\examples{
a <- get_dummy_data(num_items=60)
}
