% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReliability.R
\name{getReliability}
\alias{getReliability}
\title{getReliability}
\usage{
getReliability(tests)
}
\arguments{
\item{tests}{Vector of test names.}
}
\description{
This function extracts reliabilities from xxx_shw.xls files and puts them into
a dataframe.
}
