% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTest.R
\name{getTest}
\alias{getTest}
\title{getTest}
\usage{
getTest(run, tests, intLst)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{tests}{Vector of test names.}

\item{intLst}{List of item analysis dataframe for each test.}
}
\description{
This function extracts reliabilities from xxx_shw.xls files and puts them into
a dataframe.
}
\examples{
# Not run
# getTest(tests=c('V', 'Q'))
}
