% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo_effect.R
\name{halo_effect}
\alias{halo_effect}
\title{halo_effect}
\usage{
halo_effect(df, no_dim)
}
\arguments{
\item{df}{Dataframe with only double marks on scoring criteria. The first 
half of columns are the first marks on each criterion, and the second half 
the second marks.}

\item{no_dim}{Number of scoring criteria.}
}
\value{
Dataframe of indices to flag halo effect. The default threshold is 1.2.
}
\description{
This function calculates halo effect indexes for marks on constructed responses.
}
\examples{
halo_effect()
}
