% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages_ls.R
\name{install_packages_ls}
\alias{install_packages_ls}
\title{install_packages_ls}
\usage{
install_packages_ls(
  packages = c("plyr", "conquestr", "rlang", "ggthemes", "ggrepel", "patchwork", "gdata",
    "janitor", "data.table", "RColorBrewer", "fs", "here", "writexl", "ggpubr", "qpdf",
    "tidyverse")
)
}
\arguments{
\item{packages}{Vector of package names (character). Default is the vector of packages needed for Package RaschKit.}
}
\description{
This function installs packages if not yet installed.
}
\examples{
install_packages_ls(packages='openxlsx')
}
