% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_stats.R
\name{item_stats}
\alias{item_stats}
\title{item_stats}
\usage{
item_stats(run, test)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of test.}
}
\description{
This function extracts item statistics such as item-rest correlation,
facility, delta, and case number from 'xxx_its.txt' and 'xxx_shw.txt' files
and merge them into one dataframe. This is associated
with test named 'test'.
}
\examples{
a <- item_stats(folder='output', test='Hum', anchor=F)
}
