% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itn_ls.R
\name{itn_ls}
\alias{itn_ls}
\title{itn_ls}
\usage{
itn_ls(run, test)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of test.}
}
\description{
This function extract item analysis results for each item from xxx_itn.txt file
and put them into a list. This is associated with test named 'test'.
}
