% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itn_poly_comment.R
\name{itn_poly_comment}
\alias{itn_poly_comment}
\title{itn_poly_comment}
\usage{
itn_poly_comment(run, test)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of test.}
}
\description{
This function checks whether polytomous items' neighbor categories' ability
estimates are too small. This is associated with test named 'test'.
}
\examples{
itn_poly_comment(run, test)
}
