% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnFlag123.R
\name{itnFlag123}
\alias{itnFlag123}
\title{itnFlag123}
\usage{
itnFlag123(
  run,
  test,
  easy = 90,
  hard = 10,
  iRst = 0.11,
  fit_w = 1.1,
  fit_uw = 1.2
)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of test.}

\item{easy}{Threshold to flag easy items. Default is 90 (percent correct).}

\item{hard}{Threshold to flag hard items. Default is 10 (percent correct).}

\item{iRst}{Threshold to flag low item-rest correlation statistics. Default is 0.11.}

\item{fit_w}{Threshold to flag large weighted item fit statistics. Default is 1.1.}

\item{fit_uw}{Threshold to flag large unweighted item fit statistics. Default is 1.2.}
}
\value{
Dataframe of item statistics with flags.
}
\description{
This function adds three priorities of flag to items in a test based on item
statistics. This is associated with test named 'test'.
}
\examples{
a <- itnFlag123(test='math_35')
}
