% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_into_Data.R
\name{keys_into_Data}
\alias{keys_into_Data}
\title{keys_into_Data}
\usage{
keys_into_Data(test, keys)
}
\arguments{
\item{test}{Name of test.}

\item{keys}{Vector of item keys in the test, e.g., c('A', 'C', 'D')}
}
\description{
This function saves raw key file 'test_Key.txt' into 'data' folder. Later, 
this saved file is used to produce the key section of 'test.cqc' file in 
'input' folder. This is associated with test named 'test'.
}
\examples{
keys_into_Data()
}
