% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lines.R
\name{Lines}
\alias{Lines}
\title{Lines}
\usage{
Lines(folder, test, type, str)
}
\arguments{
\item{folder}{Place of output files from ConQuest. Default is NULL where
'output' folder is used.}

\item{test}{Name of test.}
}
\description{
This function detects lines where 'str' appears in file of 'type' and 'test'.
}
