% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missAll_remove_cols.R
\name{missAll_remove_cols}
\alias{missAll_remove_cols}
\title{missAll_remove_cols}
\usage{
missAll_remove_cols(df)
}
\arguments{
\item{df}{Dataframe with columns of all missing values.}
}
\value{
Dataframe without columns of all missing values.
}
\description{
This function removes columns of all missing values from a dataframe.
}
\examples{
missAll_remove_cols()
}
