% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missAll_remove_rows.R
\name{missAll_remove_rows}
\alias{missAll_remove_rows}
\title{missAll_remove_rows}
\usage{
missAll_remove_rows(df, begin, end, rm = TRUE)
}
\arguments{
\item{df}{Dataframe that contains responses of all missing values.}

\item{begin}{The 1st item's order in test, e.g., 7.}

\item{end}{The last item order in test, e.g., 46.}

\item{rm}{Whether to remove all-missing cases. Default is TRUE.}
}
\value{
Dataframe that has responses of all missing values removed.
}
\description{
This function removes cases with all missing responses of
c('@', '@@', NA, '9', '.').
}
\examples{
missAll_remove_rows()
}
