% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_into_folder.R
\name{move_into_folder}
\alias{move_into_folder}
\title{move_into_folder}
\usage{
move_into_folder(folder = "output", test)
}
\arguments{
\item{folder}{Place to find CQ output files. Default is 'output' folder.}

\item{test}{Name of the test.}
}
\description{
This function checks whether files related to 'test' exist. If so,
it establishes a new folder and moves files inside. This is associated
with test named 'test'.
}
\examples{
move_into_folder(test=test)
}
