% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_item.R
\name{N_item}
\alias{N_item}
\title{N_item}
\usage{
N_item(folder, test)
}
\arguments{
\item{folder}{Place of output files from ConQuest. Default is NULL where
'output' folder is used.}

\item{test}{Name of test.}
}
\description{
This function finds out number of items in a test. This is associated
with test named 'test'.
}
