% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_item_Step.R
\name{N_item_Step}
\alias{N_item_Step}
\title{N_item_Step}
\usage{
N_item_Step(folder, test)
}
\arguments{
\item{folder}{Folder where .shw file is located.}

\item{test}{Name of test.}
}
\description{
This function extracts line numbers from .shw file and calculates number of item steps. This is associated with test named 'test'.
}
\examples{
N_item_Step()
}
