% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_stats.R
\name{opt_stats}
\alias{opt_stats}
\title{opt_stats}
\usage{
opt_stats(run, test)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of test.}
}
\value{
Dataframe of option statistics.
}
\description{
This function extracts option statistics from 'test.txt' file.
This is associated with test named 'test'.
}
\examples{
b <- opt_stats(test='math_35')
}
