% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pid_resp_regrs_cols.R
\name{pid_resp_regrs_cols}
\alias{pid_resp_regrs_cols}
\title{pid_resp_regrs_cols}
\usage{
pid_resp_regrs_cols(
  df,
  pid,
  n_cov,
  n_resp,
  regr_vec = NULL,
  pweight = NULL,
  group = NULL
)
}
\arguments{
\item{df}{Dataframe with ID, covariates and responses.}

\item{pid}{Variable name of person ID.}

\item{n_cov}{Number of covariates before responses.}

\item{n_resp}{Number of responses.}

\item{regr_vec}{String vector of regressors' names. Default is NULL.}

\item{group}{Either NULL or variable name in data, for group plotting.}
}
\value{
e.g., list(pid_cols='14-30', resps_cols='35-308', regr_ls=list(...))
}
\description{
This function calculates and outputs some variables' beginning and ending
in a dataframe if saved as a fixed-width .txt file. Those variables include
person ID, responses, and perhaps regressors.
}
\examples{
pid_resp_regrs_cols(df=cov_respns, pid='pid', n_cov=5, n_resp=2)
pid_resp_regrs_cols(df=cov_respns, pid='pid', n_cov=5, n_resp=2,
regr_vec=c('nation', 'gender', 'G2'))
}
