% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DIF.R
\name{plot_DIF}
\alias{plot_DIF}
\title{plot_DIF}
\usage{
plot_DIF(
  df,
  wh,
  vars,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  step = FALSE,
  DIFVar = NULL,
  cor,
  shift,
  sdr,
  axRange,
  quick = TRUE
)
}
\arguments{
\item{df}{Dataframe with results of chi-square tests on items that appeared
in two tests.}

\item{wh}{'Before' or 'After' when to plot before or after equating or in
DIF analysis.}

\item{vars}{Vector of length 2 such as c('VIC','NSW'). Its order corresponds
to .x and .y (delta, error).}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two
tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference
between two tests. Default is 4.}

\item{step}{TRUE if DIF analysis is performed on step parameters. Default is FALSE.}

\item{DIFVar}{Name of dichotomous DIF variable, e.g., 'gender'. Default is NULL.}

\item{cor}{Correlation of two categories' delta estimates.}

\item{shift}{Shift of two categories' average delta estimates.}

\item{sdr}{SD ratio of two categories' delta estimates.}

\item{axRange}{Vector of range of x and y axis.}

\item{quick}{Whether quick error is needed. Default is TRUE for DIF analysis.}
}
\value{
Plot of DIF analysis.
}
\description{
This function plots all items in a test or all anchors and label DIF items/anchors.
}
