% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DIF_group.R
\name{plot_DIF_group}
\alias{plot_DIF_group}
\title{plot_DIF_group}
\usage{
plot_DIF_group(test, DIFVar, numAbilGrps = 3, abilEst2use = "pv1")
}
\arguments{
\item{test}{Name of test.}

\item{DIFVar}{Name of DIF variable. Default is NULL.}

\item{numAbilGrps}{Number of ability groups. Default is 3.}

\item{abilEst2use}{Ability type used for curve data. Default is 'pv1'.
For smaller sample size,  'wle' should be considered.}
}
\description{
This function plots expected score curves for each item using results from
item calibration with the DIF variable as a group variable. This is associated
with test named 'test'.
}
\examples{
plot_DIF_group(test='RANZCOG', DIFVar='Educator')
}
