% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facil.R
\name{plot_facil}
\alias{plot_facil}
\title{plot_facil}
\usage{
plot_facil(test, facil_dif, vars, facil_cut, steps, DIFVar)
}
\arguments{
\item{test}{Name of test.}

\item{facil_dif}{Dataframe of two categories' facilities, output from function facil_DIF().}

\item{vars}{Vector of two categories' names, e.g., c('girls', 'boys').}

\item{facil_cut}{Threshold to label items with big facility difference between the two categories.}

\item{steps}{Whether polytomous items are involved.}
}
\value{
Plot of one category's facilities vs. those of the other category of the dichotomous DIF variable.
}
\description{
This function plots the facilities of the two categories of the dichotomous DIF variable.
}
\examples{
plot_facil()
}
