% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facilDiscrFitw.R
\name{plot_facilDiscrFitw}
\alias{plot_facilDiscrFitw}
\title{plot_facilDiscrFitw}
\usage{
plot_facilDiscrFitw(facilDiscrFitw, tests, colsInd1, numInd)
}
\arguments{
\item{facilDiscrFitw}{Dataframe of two tests' statistics from its.txt files.}

\item{tests}{Vector of test names in output files to be used to locate files.}

\item{colsInd1}{Two tests' column numbers of 1st index in dataframe.}

\item{numInd}{Number of index.}
}
\description{
This function plots the scatterplots of facilities, discrimination, and
weighted fit of two tests.
}
