% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_int_slop.R
\name{plot_int_slop}
\alias{plot_int_slop}
\title{plot_int_slop}
\usage{
plot_int_slop(test, x, y)
}
\arguments{
\item{test}{Name of test.}

\item{x}{Vector of numbers that correspond to x-axis.}

\item{y}{Vector of numbers that correspond to y-axis.}
}
\value{
Scatter plot with annotation information.
}
\description{
This function draws a scatter plot for two vectors of numbers. This is 
associated with test named 'test'. The plot also shows the formula, 
squared error explained, and RMSE.
}
\examples{
plot_int_slop()
}
