% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_recode.R
\name{poly_recode}
\alias{poly_recode}
\title{poly_recode}
\usage{
poly_recode(
  keys,
  data,
  n_cov,
  miss_code = c("r", "R", "m", "M", "9", "x", "X", ".", "", " ", NA)
)
}
\arguments{
\item{keys}{Vector of keys in the test. Default is NULL.}

\item{data}{Dataframe with pid, covariables (e.g,, DIF variable), and
responses. Default is NULL where Excel file with name 'test' in 'data'
folder is used.}

\item{n_cov}{Number of covariates before responses.}

\item{miss_code}{Missing codes. Default is c('r','R','m','M','9','x','X','.','',' ',NA).}
}
\description{
This function down ranks non-continuous polytomous item response vectors
in a dataframe.
}
