% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_recode_one.R
\name{poly_recode_one}
\alias{poly_recode_one}
\title{poly_recode_one}
\usage{
poly_recode_one(
  x0,
  miss_code = c("r", "R", "m", "M", "9", "x", "X", ".", "", " ", NA)
)
}
\arguments{
\item{x0}{A vector of non-continuous polytomous item response.}

\item{miss_code}{Missing codes. Default is c('r','R','m','M','9','x','X','.','',' ',NA).}
}
\description{
This function down codes a non-continuous polytomous item response to be continuous and starting from 0. It will keep intact many possible missing codes including c('r','R','m','M','9','x','X','.','',' ',NA).
}
\examples{
poly_recode_one(c(0,10,2,2,4,6,10))
poly_recode_one(c(0,1,2,3))
poly_recode_one(c(0,'.',10,'r','R',2,2,'M','9','x',4,6,10,'m','X',' '))
poly_recode_one(c(0,10,7,8,2,2,'M',9,4,6,10,'r'), miss_code=c(7,8,9,'M','r'))
}
