% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA2Df.R
\name{QA2Df}
\alias{QA2Df}
\title{QA2Df}
\usage{
QA2Df(id, charVarVec = NULL, numVarVec = NULL, selfDf, otherDf)
}
\arguments{
\item{id}{Unique ID column in both dataframes. This should be all same across
the two dataframes.}

\item{charVarVec}{Vector of char variables both dataframes contain. Equivalence
will be checked. Default is NULL.}

\item{numVarVec}{Vector of numeric variables both dataframes contain. Difference of
each variable in this vector will be calculated. Default is NULL.}

\item{selfDf}{Dataframe of your own.}

\item{otherDf}{Dataframe of others'.}
}
\description{
This function checks whether two dataframes have same ID column and calculates
difference of each variable in the given name vector.
}
