% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qid_its.R
\name{qid_its}
\alias{qid_its}
\title{qid_its}
\usage{
qid_its(folder, test)
}
\arguments{
\item{folder}{Folder where ConQuest output files are located.}

\item{test}{Name of test.}
}
\value{
Vector of item orders.
}
\description{
This function generates vector of item orders that appeared in _its.txt file.
This is associated with test named 'test'.
}
\examples{
qid_its('output', 'FPA')
}
