% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2one.R
\name{read2one}
\alias{read2one}
\title{read2one}
\usage{
read2one(
  folder = c("results", "DIF", "equating"),
  tests,
  prefix = NULL,
  file_name = NULL,
  vars = NULL
)
}
\arguments{
\item{folder}{Folder where to-be-read files are located. Must be one of
c('results', 'DIF', 'equating').}

\item{tests}{Vector of test names after '_'.}

\item{prefix}{Common prefix in file names before first '_'. Default is NULL.}

\item{file_name}{Name given to new file, each sheet of which is one test's result.
Default is NULL.}

\item{vars}{Vector of length 2 such as c('girls','boys'). Default is NULL.
Should be specified when to read multiple DIF analysis (binary DIF variable)
Excels and plots from 'DIF' folder.
Its order corresponds to the alphabetic/numeric order of DIF variables'
two categories in data.}
}
\description{
This function reads from 'folder' files associated with 'tests'
and puts them into one file.
}
\examples{
# Not run
# read2one(folder='equating', tests=c('AHU','MST','R','N'))
}
