% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_keys.R
\name{read_keys}
\alias{read_keys}
\title{read_keys}
\usage{
read_keys(folder, test)
}
\arguments{
\item{folder}{The 'data' folder where key 'test_Key.txt' file is located.}

\item{test}{Name of the test.}
}
\description{
This function read from 'data' folder the key file associated with 'test'.
}
\details{
They key file has one row for all-multiple-choice-item test, or multiple 
rows if any item has double key or polytomous scoring.
}
