% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resps_modify.R
\name{resps_modify}
\alias{resps_modify}
\title{resps_modify}
\usage{
resps_modify(keyDf, resps_cols, delVec)
}
\arguments{
\item{keyDf}{Dataframe of 'Item', 'Key', and 'Max_score' (add Key2 if double key).}

\item{resps_cols}{Column number(s) of responses in the data .txt file.}

\item{delVec}{Vector of item order number to be removed from the test,
e.g., c(2, 3, 45, 46).}
}
\description{
This function consecutively remove the last item from the delete list of
items until no last item is in it.
}
