% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_dim.R
\name{section_dim}
\alias{section_dim}
\title{section_dim}
\usage{
section_dim(scrs, nDimVec, dimNmVec)
}
\arguments{
\item{scrs}{Vector of possible scores in the test.}

\item{nDimVec}{Vector of numbers of responses the dimensions have.}

\item{dimNmVec}{Vector of the dimensions' names.}
}
\description{
This function generates the dimension and relevant label sessions of
the .cqc file when the test has more than one dimension.
}
\examples{
section_dim(codes=c(1, 2, 3, 4), nDimVec=c(30, 45),
dimNmVec=c('Literacy', 'Numeracy'))
}
