% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_estimate.R
\name{section_estimate}
\alias{section_estimate}
\title{section_estimate}
\usage{
section_estimate(quick)
}
\arguments{
\item{quick}{TRUE if quick estimation is preferred.}
}
\value{
String of characters used in estimation section of 'test.cqc' file in 'input' folder.
}
\description{
This function creates string of estimation specification for 'test.cqc' file in 'input' folder. This is associated with test named 'test'.
}
\examples{
section_estimate()
}
