% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_keys.R
\name{section_keys}
\alias{section_keys}
\title{section_keys}
\usage{
section_keys(folder, test, dbl_key, poly_key, delete)
}
\arguments{
\item{folder}{The 'data' folder where key 'test_Key.txt' file is located.}

\item{test}{Name of the test.}

\item{dbl_key}{TRUE if the key of any item is double key.}

\item{poly_key}{TRUE if the key of any item has polytomous scoring.}

\item{delete}{Vector of to-be-removed items' order numbers in the test, e.g., c(4, 7, 65, 114).}
}
\value{
String of characters used in key section of 'test.cqc' file in 'input' folder.
}
\description{
This function reads in 'test_Key.txt' from 'data' folder and specifies the Key section of the 'test.cqc' file in 'input' folder. Both procedures are associated with test named 'test'.
}
\details{
They key file has one row for all-multiple-choice-item test, or multiple rows if any item has double key or polytomous scoring.
}
\examples{
section_keys()
}
