% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_model.R
\name{section_model}
\alias{section_model}
\title{section_model}
\usage{
section_model(regr_ls, codes, poly_key, DIFVar, step, poly_group)
}
\arguments{
\item{regr_ls}{List of regressors with reference group excluded,
e.g., list('G4'='110', 'G5'='111', 'G6'='112').}

\item{codes}{Vector of valid codes for item responses, e.g., c(1, 2, 3, 4, 5, 6, 7, 8, 9).}

\item{poly_key}{TRUE if any item has polytomous scoring.}

\item{DIFVar}{Name of DIF variable.}

\item{step}{TRUE if DIF analysis is performed on step parameters.
Default is FALSE.}

\item{poly_group}{TRUE if model is run per group and no DIF variable is used.}
}
\value{
String of characters used in model section of 'test.cqc' file in
'input' folder.
}
\description{
This function creates string of model specification for 'test.cqc' file
in 'input' folder. This is associated with test named 'test'.
}
\examples{
section_model()
}
