% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_specs.R
\name{section_specs}
\alias{section_specs}
\title{section_specs}
\usage{
section_specs(anchor, test, DIFVar, poly_catgrs, quick)
}
\arguments{
\item{anchor}{TRUE when anchor is to be done.}

\item{test}{Name of the test.}

\item{DIFVar}{Name of the DIF variable.}

\item{poly_catgrs}{Vector of categories of the polytomous DIF variable 'DIFVar'.}

\item{quick}{TRUE when testing.}
}
\value{
String of characters used in estimation section of 'test.cqc'
file in 'input' folder.
}
\description{
This function specifies the section of estimation details in 'test.cqc'
file in 'input' folder. This is associated with test named 'test'.
}
\details{
When anchoring is to be done, the import route of the anchor file is
designated and item constraint is specified as 'constraints=none'.
When DIF analysis is to be done on a polytomous DIF variable, it is
specified that some category of the DIF variable is to be selected from
among all of its categories.
}
\examples{
section_specs()
}
