% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_its.R
\name{skip_its}
\alias{skip_its}
\title{skip_its}
\usage{
skip_its(file)
}
\arguments{
\item{file}{File name to read.}
}
\value{
Number of lines to skip.
}
\description{
This function finds out the number of lines to skip when reading 'test.its' 
output file from ConQuest.
}
\examples{
skip_its()
}
