% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_file.R
\name{str_file}
\alias{str_file}
\title{str_file}
\usage{
str_file(folder, test, type)
}
\arguments{
\item{folder}{Folder where ConQuest output files are located.}

\item{test}{Name of test.}

\item{type}{File type.}
}
\value{
Vector of item orders.
}
\description{
This function reads into string a file of 'type' and 'test'.
This is associated with test named 'test'.
}
\examples{
str_file('FPA', 'shw')
}
