% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term3_L2.R
\name{term3_L2}
\alias{term3_L2}
\title{term3_L2}
\usage{
term3_L2(folder, test)
}
\arguments{
\item{folder}{Folder where .shw file is located.}

\item{test}{Name of test.}
}
\description{
This function finds out a line number. This is associated with test named 'test'.
}
\examples{
term3_L2()
}
