% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_L2.R
\name{term_L2}
\alias{term_L2}
\title{term_L2}
\usage{
term_L2(folder, test)
}
\arguments{
\item{folder}{Folder where .shw file is located.}

\item{test}{Name of test.}
}
\description{
This function finds out a line number. This is associated with test named 'test'.
}
