% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_L2_Step.R
\name{term_L2_Step}
\alias{term_L2_Step}
\title{term_L2_Step}
\usage{
term_L2_Step(folder, test)
}
\arguments{
\item{folder}{Folder where .shw file is located.}

\item{test}{Name of test.}
}
\description{
This function a line number close to interaction term estimates of 'item*step' in .shw file. This is associated with test named 'test'.
}
