% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_cols.R
\name{var_cols}
\alias{var_cols}
\title{var_cols}
\usage{
var_cols(df, var_name)
}
\arguments{
\item{df}{Dataframe with ID, covariates and responses.}

\item{var_name}{Variable name in the dataframe, e.g., 'gender'.}
}
\value{
List. Name is variable. Element is variable's column numbers in
'test_Data.txt' file in 'data' folder.
}
\description{
This function calculates and outputs a variable's beginning and ending in
a dataframe if saved as a fixed-width .txt file.
}
\examples{
var_cols(resp, "ID number")
}
