% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fwf.R
\name{write_fwf}
\alias{write_fwf}
\title{write_fwf}
\usage{
write_fwf(dat, file)
}
\arguments{
\item{dat}{Dataframe which is usually read from anchor file in 'output' folder.}

\item{file}{Path of processed anchor file to be saved, usually in 'input' folder.}
}
\description{
This function inserts seven spaces between columns of dataframe 'dat' and
write the final string to a text file with path 'file'. This is usually
used to process anchor files.
}
