% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_process.R
\name{anchor_process}
\alias{anchor_process}
\title{anchor_process}
\usage{
anchor_process(test, respDf, keyDf, n_cov, nDimVec, ancDf)
}
\arguments{
\item{test}{Name of test.}

\item{respDf}{Dataframe with pid, covariables (e.g,, DIF variable), and responses.
Default is NULL where Excel file with name 'test' in 'data' folder is used.}

\item{keyDf}{Vector of keys in the test. Default is NULL.}

\item{n_cov}{Number of covariates before responses.}

\item{nDimVec}{Vector of numbers of responses the dimensions have. Default is NULL.
Define this vector if multi-dimensional model is to be run, e.g., c(30, 45).
Also should define this if there are variables after response columns, e.g., 30.}

\item{ancDf}{Dataframe of 'Item' and 'Delta' for anchors.}
}
\description{
This function removes deleted or no-data items, reorders items,
extracts anchors with new orders, and saves file into 'input' folder. The file
will be imported into ConQuest as anchor file. This is
associated with test named 'test'.
}
