% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{calibrate}
\usage{
calibrate(
  wd = here::here(),
  folder = here::here("output"),
  test,
  data = NULL,
  pid,
  n_cov,
  regr_vec_char = NULL,
  n_dims = NULL,
  dim_names = NULL,
  keys = NULL,
  labels = NULL,
  quick = FALSE,
  delete = NULL,
  dbl_key = FALSE,
  poly_key = FALSE,
  anchor = FALSE,
  section_extr = NULL,
  easy = 90,
  hard = 10,
  iRst = 0.11,
  fit_w = 1.1,
  fit_uw = 1.2,
  dFallThr = 0.5,
  dRiseThr = 0.1,
  sav_xlsx = TRUE,
  abilEst2use = "pv1",
  numAbilGrps = NULL,
  recode = TRUE,
  long_label = FALSE
)
}
\arguments{
\item{wd}{Working directory. Default is the folder where .Rproj is located.}

\item{folder}{Place to find CQ output files.}

\item{test}{Name of the test.}

\item{data}{Dataframe with pid, covariables (e.g,, DIF variable), and responses. Default is NULL where Excel file with name 'test' in 'data' folder is used.}

\item{pid}{Name of candidates' ID variable.}

\item{n_cov}{Number of covariates before responses.}

\item{regr_vec_char}{Vector of character regressors' names.}

\item{n_dims}{Vector of numbers of responses the dimensions have. Default is NULL. Define this vector if multi-dimensional model is to be run, e.g., c(30, 45). Also should define this if there are variables after response columns, e.g., 30.}

\item{dim_names}{Vector of the dimensions' names. Default is NULL. Define this vector if multi-dimensional model is to be run.}

\item{keys}{Vector of keys in the test. Default is NULL.}

\item{labels}{Vector of item labels that correspond to order of item response columns in data.}

\item{quick}{TRUE when testing. Default is FALSE.}

\item{delete}{Vector of orders of items to be removed. Default is NULL.}

\item{dbl_key}{TRUE if any item has polytomous scoring. Default is FALSE.}

\item{poly_key}{TRUE if the key of any item has polytomous scoring. Default is FALSE.}

\item{anchor}{TRUE when anchor is to be done. Default is FALSE.}

\item{section_extr}{Extra sections to be added to 'test.cqc' file in 'input' folder. Default is NULL.}

\item{easy}{Threshold to flag easy items. Default is 90 (means 90 percent correct).}

\item{hard}{Threshold to flag hard items. Default is 10 (means 10 percent correct).}

\item{iRst}{Threshold to flag low item-rest correlation statistics. Default is 0.11.}

\item{fit_w}{Threshold to flag large weighted item fit statistics. Default is 1.1.}

\item{fit_uw}{Threshold to flag large unweighted item fit statistics. Default is 1.2.}

\item{dFallThr}{Ability on last bin above which falling distractor is flagged. Default is 0.5.}

\item{dRiseThr}{Ability on last bin below which rising distractor is unflagged. Default is 0.1.}

\item{sav_xlsx}{Whether to save Excel file to 'results' folder. Default is TRUE.}

\item{abilEst2use}{Ability type used for curve data. Default is 'pv1'. Use 'wle' for smaller samples.}

\item{numAbilGrps}{Number of ability groups. Default is NULL.}

\item{long_label}{Whether item labels are longer than 16 characters' fixed width. Default is FALSE.}

\item{step}{TRUE if polytomous items are involved. Default is FALSE.}
}
\description{
This function calibrates items in a test and adds comments and flags of levels of priority for review. This is associated with test named 'test'. If sav_xlsx is TRUE, an Excel file with a summary of the above information will be saved in 'results' folder in the working directory.
}
\examples{
calibrate(data=cov_respns, test='randomData_Dim3', pid="pid", n_cov=6, n_dims=c(20, 10), regr_vec_char=c('nation', 'age', 'language', 'grade'), dim_names=c('English', 'Maths'), quick=TRUE, all_mc=FALSE, dim_multi=TRUE)
calibrate(data=elena, test='elena_Dich2', pid="IDSTUD", n_cov=9, quick=TRUE, delete=c(1,6,8,9,28,34,35,36))
calibrate(data=elena, test='elana_math1', pid="IDSTUD", n_cov=9, quick=TRUE, all_mc=FALSE, poly_key=TRUE)
calibrate(data=racp, test='racp1', pid="V1", n_cov=1, quick=TRUE, keys=cd$`Correct options`, labels=cd$`Question ID`)
calibrate(data=elena, test='elana_poly_score', pid="IDSTUD", n_cov=9, quick=TRUE, poly_key=TRUE)
}
