% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{calibrate}
\usage{
calibrate(
  test,
  data = NULL,
  keys,
  pid,
  n_cov,
  regr_vec_char = NULL,
  n_dims = NULL,
  dim_names = NULL,
  quick = TRUE,
  delete = NULL,
  dbl_key = NULL,
  anchor = FALSE,
  section_extr = NULL,
  easy = 90,
  hard = 10,
  iRst = 0.11,
  fit_w = 1.1,
  fit_uw = 1.2,
  dFallThr = 0.5,
  dRiseThr = 0.1,
  numAbilGrps = NULL,
  recode_poly = FALSE,
  missCode2Conv = NULL,
  filetype = "sav",
  slope = NULL,
  intercept = NULL,
  extrapolation = FALSE,
  save_xlsx = TRUE,
  est_type = NULL,
  sparse_check = FALSE,
  CCCip2Wd = FALSE,
  pweight = NULL,
  dfAnc = NULL
)
}
\arguments{
\item{test}{Name of the test.}

\item{data}{Dataframe with pid, covariables (e.g,, DIF variable), and
responses. Default is NULL where Excel file with name 'test' in 'data'
folder is used.}

\item{keys}{Dataframe of 'Item', 'Key', and 'Max_score' (add Key2 if double key).}

\item{pid}{Name of candidates' ID variable.}

\item{n_cov}{Number of covariates before responses.}

\item{regr_vec_char}{Vector of character regressors' names.}

\item{n_dims}{Vector of numbers of responses the dimensions have.
Default is NULL. Define this vector if multi-dimensional model is to be run,
e.g., c(30, 45). Also should define this if there are variables after
response columns, e.g., 30.}

\item{dim_names}{Vector of the dimensions' names. Default is NULL.
Define this vector if multi-dimensional model is to be run.}

\item{quick}{TRUE if empirical error is not needed. Default is TRUE}

\item{delete}{Vector of orders or labels of items to be removed. Default is NULL.}

\item{dbl_key}{TRUE if any item has polytomous scoring. Default is NULL.}

\item{anchor}{TRUE when anchor is to be done. Default is FALSE.}

\item{section_extr}{Extra sections to be added to 'test.cqc' file in
'input' folder. Default is NULL.}

\item{easy}{Threshold to flag easy items. Default is 90 (percent correct).}

\item{hard}{Threshold to flag hard items. Default is 10 (percent correct).}

\item{iRst}{Threshold to flag low item-rest correlation statistics.
Default is 0.11.}

\item{fit_w}{Threshold to flag large weighted item fit statistics.
Default is 1.1.}

\item{fit_uw}{Threshold to flag large unweighted item fit statistics.
Default is 1.2.}

\item{dFallThr}{Ability on last bin above which falling distractor is flagged.
Default is 0.5.}

\item{dRiseThr}{Ability on last bin below which rising distractor is unflagged.
Default is 0.1.}

\item{numAbilGrps}{Number of ability groups. Default is NULL.}

\item{recode_poly}{TRUE if polytomous items have non-continuous scores.
Note that code in c('r','R','m','M','9','x','X','.','',' ',NA) will be kept
intact while other codes will be down scored to be continuous. Default is FALSE.}

\item{missCode2Conv}{Response codes to convert to embedded/trailing missing
('M' or 'R'). Use freq_resps_cat() to explore response distribution
and determine which codes need recoding. Commonly found missing symbols
are '@' (less or more), 7, 8, 9, 88, 99, '.', '-', '', NA. Default is NULL
when responses have been recoded.}

\item{filetype}{Format for input dataset. Default is 'sav'. Also support
csv format.}

\item{slope}{Slope to multiply ability estimates. Default is NULL}

\item{intercept}{Value/intercept to add to ability estimates. Default is NULL.}

\item{extrapolation}{Whether to extrapolate the minimum and maximum estimates.
Default is FALSE.}

\item{save_xlsx}{Whether to save summary file. Default is TRUE (single test).}

\item{est_type}{Type of ability estimate to use for score equivalence table,
'wle' or 'mle'. 'wle' is commonly used. Default is NULL (equiv table not needed).}

\item{sparse_check}{Whether to check response column sparsity in general or
regarding any DIF variable category. Default is FALSE. If TRUE, sparse
response columns will be removed.}

\item{CCCip2Wd}{TRUE if to save CCC and item-person map to a Word file. Default
is FALSE.}

\item{pweight}{Variable name of person weights in response dataframe. Should
be specified if weight is used for modeling. Default is NULL.}

\item{dfAnc}{Dataframe of 'Item' and 'Delta' for anchors. Default is NULL.
If NULL, an xxx_anc.txt file with anchor tbl (Item, Delta) should be put
in 'input' folder beforehand. Check output 'xxx_anc.txt' file from previous run
for correct anchor order, especially for polytomous items with step parameters.}
}
\description{
This function calibrates items in a test and adds comments and flags of
levels of priority for review. This is associated with test named 'test'.
If save_xlsx is TRUE, an Excel file with a summary of the above information
will be saved in 'results' folder in the working directory.
}
\examples{
# Not run
# calibrate(data=racp, test='RACP', pid="V1", n_cov=1, keys=cd,
# delete=c(3,4,5,36), dbl_key=list(`7`=c(1,3), `9`=c(3,4)))
}
