% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_concurrent.R
\name{calibrate_concurrent}
\alias{calibrate_concurrent}
\title{calibrate_concurrent}
\usage{
calibrate_concurrent()
}
\arguments{
\item{test}{Name of test.}
}
\description{
This function combines all steps after T. This is associated with test named 'test'.
}
\examples{
calibrate_concurrent()
}
