% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateLst.R
\name{calibrateLst}
\alias{calibrateLst}
\title{calibrateLst}
\usage{
calibrateLst(
  respDfLst = NULL,
  keyDfLst,
  pid,
  n_cov,
  run,
  useR = FALSE,
  quick = TRUE
)
}
\arguments{
\item{respDfLst}{List of response dataframe. Default is NULL (reruns after review).
If NULL and items need collapsing, put processed data with test name in
'data' folder before running.}

\item{keyDfLst}{List of key dataframes. Element names are test names.}

\item{pid}{Name of candidates' ID variable.}

\item{n_cov}{Number of covariates before responses.}

\item{run}{Integer that indicates round of run.}

\item{useR}{TRUE when code 'R' is used for scoring. Default is FALSE.}

\item{quick}{TRUE if empirical error is not needed. Default is TRUE.}
}
\description{
This function calibrates a list of tests and put results into a summary file
in 'results' folder with 'itn' and run number in name. Make sure elements in keyDfLst
and respDfLst (if not NULL) in same order correspond to one test.
Also, keyDfLst should have test name for each element.
}
