% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateScale.R
\name{calibrateScale}
\alias{calibrateScale}
\title{calibrateScale}
\usage{
calibrateScale(
  keyDfLst,
  pid,
  n_cov,
  est_type = "wle",
  trial = FALSE,
  ancShiftLst = NULL,
  ancTest2ReadLst = NULL,
  ancDfLst = NULL,
  ancRead = FALSE,
  useR = FALSE,
  slope = NULL,
  intercept = NULL,
  extrapolation = FALSE
)
}
\arguments{
\item{keyDfLst}{List of key dataframes. Element names are test names.}

\item{pid}{Name of candidates' ID variable.}

\item{n_cov}{Number of covariates before responses.}

\item{est_type}{Type of ability estimate to use for score equivalence table,
'wle' or 'mle'. 'wle' is commonly used. Default is 'wle'.}

\item{trial}{TRUE when trial item diagnostics is needed after anchoring is done.
Default is FALSE.}

\item{ancShiftLst}{List of shift of mean anchor delta from previous cycle obtained from
equating analysis. It will be added to each item delta of 'test' in 'output'
folder and put in 'input' folder as anchor file for 'test'. Default is NULL.}

\item{ancTest2ReadLst}{List of name of test in 'output' folder to read delta from. This
is usually concurrent calibration results such as all grades and/or all test
forms. Only delta's (and step estimates) in 'tests' are extracted and put in
'input' folder as anchor file. Default is NULL.}

\item{ancDfLst}{List of dataframe of 'Item' and 'Delta' for anchors. Include 'Step' column
for step estimates (step number). Default is NULL.
Check output 'xxx_anc.txt' file from previous run for correct anchor order,
especially for polytomous items with step parameters.}

\item{ancRead}{Whether to read anchor file in 'input' folder. Need to put anchor
file there. Default is FALSE.}

\item{useR}{TRUE when code 'R' is used for scoring. Default is FALSE.}

\item{slope}{Slope to multiply ability estimates. Default is NULL}

\item{intercept}{Value/intercept to add to ability estimates. Default is NULL.}

\item{extrapolation}{Whether to extrapolate the minimum and maximum estimates.
Default is FALSE.}
}
\description{
This function anchors a list of tests. Make sure elements in keyDfLst,
ancShiftLst (if not NULL), and ancDfLst (if not NULL) in same order
correspond to one test. Also, keyDfLst should have test name for each element.
}
