% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCC_ipMap.R
\name{CCC_ipMap}
\alias{CCC_ipMap}
\title{CCC_ipMap}
\usage{
CCC_ipMap(
  test,
  cqs,
  abilEst2use = "pv1",
  numAbilGrps = NULL,
  long_label = FALSE,
  poly_key = FALSE
)
}
\arguments{
\item{test}{Name of the test.}

\item{cqs}{CQS output file from ConQuest.}

\item{abilEst2use}{Ability type used for curve data. Default is 'pv1'.
Use 'wle' for smaller samples.}

\item{numAbilGrps}{Number of ability groups. Default is NULL.}

\item{long_label}{Whether item labels are longer than 16 characters' fixed width.
Default is FALSE.}

\item{poly_key}{TRUE if the key of any item has polytomous scoring.
Default is FALSE.}
}
\value{
Plots of CCC by category and score.
}
\description{
This function draws for each item CCC of both category and score.
Plots will be saved in 'output' folder.
}
\examples{
plot_data <- CCC_ipMap(test='AHU', cqs=cqs)
plot_data <- CCC_ipMap(test='RACP', abilEst2use='wle')
}
