% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lab.R
\name{create_lab}
\alias{create_lab}
\title{create_lab}
\usage{
create_lab(test)
}
\arguments{
\item{test}{Name of the test.}
}
\description{
This function reads in 'test_Labels.txt' from 'data' folder and creates
'test.lab' file in 'input' folder. Both procedures are associated with
test named 'test'.
}
\details{
They key file has one row for all-multiple-choice-item test, or multiple
rows if any item has double key or polytomous scoring.
}
