% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_DIF_dich.R
\name{delta_DIF_dich}
\alias{delta_DIF_dich}
\title{delta_DIF_dich}
\usage{
delta_DIF_dich(test, DIFVar, quick = TRUE)
}
\arguments{
\item{test}{Name of test.}

\item{DIFVar}{Name of dichotomous DIF variable, e.g., 'gender'.}

\item{quick}{Whether quick error is needed. Default is TRUE for DIF analysis.}
}
\value{
List of delta and indice dataframes.
}
\description{
This function extracts delta and indice statistics for two categories of
dichotomous DIF variable.
}
\examples{
delta_DIF_dich(test='Writing', DIFVar='gender')
}
