% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_DIF_dich_step.R
\name{delta_DIF_dich_step}
\alias{delta_DIF_dich_step}
\title{delta_DIF_dich_step}
\usage{
delta_DIF_dich_step(test, DIFVar)
}
\arguments{
\item{test}{Name of test.}

\item{DIFVar}{Name of dichotomous DIF variable, e.g., 'gender'.}
}
\value{
Dataframe of variables of 'delta.x', 'delta.y'.
}
\description{
This function extracts delta statistics for two categories of dichotomous
DIF variable.
}
\examples{
delta_DIF_dich_step(test='WA', DIFVar='ATSI')
}
