% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_error_DIF_dich.R
\name{delta_error_DIF_dich}
\alias{delta_error_DIF_dich}
\title{delta_error_DIF_dich}
\usage{
delta_error_DIF_dich(folder, test, long_label = FALSE)
}
\arguments{
\item{folder}{Folder where .shw file is located with delta estimates for 
interaction term between dichotomous DIF variable and item. Default is 'DIF' 
folder in working directory.}

\item{test}{Name of test.}

\item{long_label}{Whether item labels are longer than 15 characters' fixed 
width. Default is FALSE.}
}
\value{
Dataframe of three variables of 'item', 'error.x', 'error.y'.
}
\description{
This function extracts error statistics for two categories of dichotomous 
DIF variable.
}
\examples{
delta_error_DIF_dich(folder=folder, test='RandD2')
delta_error_DIF_dich(folder=folder, test='RandD2', long_label=TRUE)
}
