% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_del.R
\name{df_del}
\alias{df_del}
\title{df_del}
\usage{
df_del(folder = here::here("output"), test, long_label = FALSE)
}
\arguments{
\item{folder}{Folder where .del file is located.}

\item{test}{Name of test.}

\item{long_label}{Whether item labels are longer than 16 characters' fixed width. Default is FALSE.}
}
\value{
Dataframe of item name_step and delta.
}
\description{
This function extracts step parameter estimates from .del file. This is associated with test named 'test'. If any item label is longer than 16 characters' fixed width, use long_label=TRUE to read complete labels from 'test_Labels.txt' file in 'data' folder.
}
\examples{
df_del(test='elana_math_NSW')
df_del(test='elana_math_NSW', long_label=TRUE)
}
