% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_del_shw_Step.R
\name{df_del_shw_Step}
\alias{df_del_shw_Step}
\title{df_del_shw_Step}
\usage{
df_del_shw_Step(folder = "output", test)
}
\arguments{
\item{folder}{Folder where xxx_shw.txt file is located. Default is 'output' folder.}

\item{test}{Name of test.}

\item{long_label}{Whether item labels are longer than 15 characters' fixed
width. Default is FALSE.}
}
\value{
Dataframe of item name_step, delta, and error.
}
\description{
This function extract item steps' delta estimates from .shw file. This is
associated with test named 'test'.
}
\examples{
df_del_shw_Step(test='elana_poly_score')
}
