% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_its.R
\name{df_its}
\alias{df_its}
\title{df_its}
\usage{
df_its(test, run)
}
\arguments{
\item{test}{Name of test.}

\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}
}
\value{
Dataframe of item statistics such as N, facility, and discrimination.
}
\description{
This function extracts parameter estimates from its.txt file. This is
associated with test named 'test'.
}
\examples{
df_its(test='Bang_3')
}
