% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_key_lab_args.R
\name{df_key_lab_args}
\alias{df_key_lab_args}
\title{df_key_lab_args}
\usage{
df_key_lab_args(
  test,
  data,
  pid,
  n_cov,
  n_resp,
  DIFVar = NULL,
  regr_vec_char = NULL,
  section_extr = NULL,
  labels = NULL,
  useR = FALSE,
  pweight = NULL
)
}
\arguments{
\item{test}{Name of the test.}

\item{data}{Dataframe with covariables, DIF variable, and responses.}

\item{pid}{Name of candidates' ID variable.}

\item{n_cov}{Number of covariates before responses.}

\item{n_resp}{Number of items in the data.}

\item{DIFVar}{DIF variable. Default is NULL.}

\item{regr_vec_char}{Vector of character regressors' names.}

\item{section_extr}{Sections to be included in the .cqc file. Default is NULL.}

\item{labels}{Item labels. Default is NULL.}

\item{useR}{TRUE when code 'R' is used for scoring. Default is FALSE.}

\item{pweight}{Variable name of person weights in response dataframe. Should
be specified if weight is used for modeling. Default is NULL.}
}
\value{
A list of arguments with calculated elements.
}
\description{
This function saves data, label, and keys into 'Data' folder in the working
directory. It also prepares the arguments needed for calibration functions.
This is associated with test named 'test'.
}
\examples{
# not run
# df_key_lab_args(data=cov_respns, test='randomData', DIFVar="age",
# keys=rep(1, 30), pid="pid", n_cov=6, n_resp=30,
# regr_vec_char=c('nation', 'language', 'grade'))
}
