% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_shw_Step.R
\name{df_shw_Step}
\alias{df_shw_Step}
\title{df_shw_Step}
\usage{
df_shw_Step(folder = "output", test)
}
\arguments{
\item{folder}{Folder where xxx_shw.txt file is located.}

\item{test}{Name of test.}
}
\description{
This function extracts step parameter estimates from .shw file. This is
associated with test named 'test'.
}
\examples{
df_shw_Step(test='elana_poly_score')
}
