% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_shw_Term3.R
\name{df_shw_Term3}
\alias{df_shw_Term3}
\title{df_shw_Term3}
\usage{
df_shw_Term3(folder, test)
}
\arguments{
\item{folder}{Folder where .shw file is located.}

\item{test}{Name of test.}
}
\description{
This function extract item statistics from 'test_facet.shw' file and add 
significance test results and flags. This is associated with test named 'test'.
}
\examples{
df_shw_Term3(folder=folder, test='randomData')
}
