% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_thr.R
\name{df_thr}
\alias{df_thr}
\title{df_thr}
\usage{
df_thr(folder = "output", test)
}
\arguments{
\item{folder}{Folder where ConQuest output files are located.}

\item{test}{Name of test.}
}
\value{
Dataframe of step parameter estimates.
}
\description{
This function extracts step parameter estimates from xxx_thr.txt file. This is
associated with test named 'test'.
}
\examples{
df_thr(test='FPA')
df_thr(test='FPA', long_label=TRUE)
}
