% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_thr.R
\name{df_thr}
\alias{df_thr}
\title{df_thr}
\usage{
df_thr(folder = "output", test, long_label = FALSE)
}
\arguments{
\item{folder}{Folder where ConQuest output files are located.}

\item{test}{Name of test.}

\item{long_label}{Whether item labels are longer than 16 characters' fixed
width. Default is FALSE.}
}
\value{
Dataframe of step parameter estimates.
}
\description{
This function extracts step parameter estimates from xxx_thr.txt file. This is
associated with test named 'test'. If any item label is longer than 16
characters' fixed width, use long_label=TRUE to read complete labels
from 'test_Labels.txt' file in 'data' folder.
}
\examples{
df_thr(test='FPA')
df_thr(test='FPA', long_label=TRUE)
}
