% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_dim_one.R
\name{DIF_dim_one}
\alias{DIF_dim_one}
\title{DIF_dim_one}
\usage{
DIF_dim_one(
  method = c("chi_square", "Bonferroni", "Facet"),
  test,
  pid,
  n_cov,
  DIFVar,
  data = NULL,
  keys = NULL,
  vars = NULL,
  poly_facet = FALSE,
  n_resp = NULL,
  regr_vec_char = NULL,
  labels = NULL,
  quick = TRUE,
  section_extr = NULL,
  miss_code = c(".", "r", "R", "x", "X", "", " "),
  poly_key = FALSE,
  prep_process = FALSE,
  save_xlsx = TRUE,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  chi_cut = 10,
  facil_cut = 10,
  iterative = FALSE,
  step = FALSE,
  desig_effect = 1
)
}
\arguments{
\item{method}{One of 'chi_square', 'Bonferroni', or 'Facet'.}

\item{test}{Name of the test.}

\item{pid}{Name of candidates' ID variable.}

\item{n_cov}{Number of covariates before responses.}

\item{DIFVar}{Name of DIF variable. Should be lowercase for ConQuest to run.
Default is NULL.}

\item{data}{Dataframe with pid, covariables (e.g,, DIF variable), and responses.
Default is NULL where Excel file with name 'test' in 'data' folder is used.}

\item{keys}{Vector of keys in the test. Default is NULL.}

\item{vars}{Vector of length 2 such as c('girls','boys'). Its order corresponds
to the alphabetic/numeric order of DIF variables' two categories in data.}

\item{poly_facet}{TRUE if facet model is to be run on a polytomous DIF variable.
Default is FALSE.}

\item{n_resp}{Number of items. Default is NULL.}

\item{regr_vec_char}{Vector of character regressors' names.}

\item{labels}{Vector of item labels that correspond to order of item response
columns in data.}

\item{quick}{Whether quick error is needed. Default is TRUE for DIF analysis.}

\item{section_extr}{Extra sections to be added to .cqc file in 'input' folder.
Default is NULL.}

\item{miss_code}{Code for missing data that should be converted to NA.
Default is c('.', 'r', 'R', 'x', 'X', '', ' ').}

\item{poly_key}{TRUE if the key of any item has polytomous scoring.
Default is FALSE.}

\item{prep_process}{TRUE if it is needed to remove items without data on
some categories of DIF variable. Default is FALSE.}

\item{save_xlsx}{Whether to save summary file and plots. Default is TRUE
(one DIF variable).}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between
two tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference
between two tests. Default is 4.}

\item{chi_cut}{Threshold of chi-square difference between two tests.
Default is 10.}

\item{facil_cut}{Threshold of number of percent to flag an item with large
facility difference between two groups of test takers. Default is 10.}

\item{iterative}{TRUE to iteratively remove DIF items. Default is FALSE}

\item{step}{TRUE if polytomous items are involved. Default is FALSE.}

\item{desig_effect}{Value to adjust errors. Default is 1.}
}
\description{
This function can perform DIF analysis on either dichotomous or polytomous
variable. This is associated with test named 'test'.
}
\details{
Should preprocess data in three ways. First, convert non-used categories and
missing values to NA. Second, remove item responses with all missing values.
Third, remove items that have all missing values or all correct/incorrect
responses on any category of the DIF variable.
}
\examples{
DIF_dim_one(method='chi_square', vars=c('Girls', 'Boys'), data=elena,
test='ELNA', DIFVar="gender", pid="IDSTUD", n_cov=9,
regr_vec_char=c('province', 'quintile'), quick=TRUE, miss_code=NULL)
DIF_dim_one(method='Bonferroni', data=elena, test='ELNA', DIFVar="quintile",
pid="IDSTUD", n_cov=9, regr_vec_char=c('province', 'gender'), quick=TRUE,
miss_code=NULL)
DIF_dim_one(method='Facet', data=elena, test='ELNA', DIFVar="quintile",
pid="IDSTUD", n_cov=9, regr_vec_char=c('province', 'gender'), quick=TRUE,
miss_code=NULL)
}
