% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_poly_shw.R
\name{DIF_poly_shw}
\alias{DIF_poly_shw}
\title{DIF_poly_shw}
\usage{
DIF_poly_shw(
  DIFVar,
  test,
  domain = NULL,
  p_cut = 0.05,
  step = FALSE,
  labels = NULL
)
}
\arguments{
\item{DIFVar}{Name of polytomous DIF variable, e.g., 'nation'.}

\item{test}{Name of the test.}

\item{domain}{Name of the domain in the test, e.g., 'Literacy'. Default is NULL.}

\item{p_cut}{Threshold of statistical test. Default is 0.05.}

\item{step}{TRUE if polytomous items are involved. Default is FALSE.}

\item{labels}{Dataframe of labels of all the items in a test. Variable
'item' is item order in label file ('test_Labels.txt' in 'data' folder)
and in test. Variable 'label' is complete item labels of test items. Default
is NULL.}
}
\value{
List of summary of results from polytomous DIF variable analysis,
including comments, step, summary statistics with flags, and statistics used
to produce the the flags.

Tibble of summary of results from polytomous DIF variable analysis.
}
\description{
This function extracts item statistics from subgroups of the polytmous
DIF variable (in 'test_category.shw' files in DIF variable's subfolder
within 'DIF' folder) to produce a list of dataframes of those statistics.
Then it calls function DIF_poly() to perform a Bonferroni-adjusted
statistical test and flag the combinations of item and category that show
significant results.
}
\examples{
DIF_poly_shw(DIFVar='educator', 'RANZCOG')
}
