% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_VarTests.R
\name{DIF_VarTests}
\alias{DIF_VarTests}
\title{DIF_VarTests}
\usage{
DIF_VarTests(
  testVec = NULL,
  respDfLst = NULL,
  difVarLst = NULL,
  n_cov = NULL,
  pid = NULL,
  keyDfLst = NULL,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  desig_effect = 1,
  step = FALSE,
  resltReady = FALSE,
  iter = TRUE,
  test3term = NULL
)
}
\arguments{
\item{testVec}{Vector of test names. Default is NULL.}

\item{respDfLst}{Dataframe of DIF variables and responses.}

\item{difVarLst}{List of DIF variables. Name is variable name. Element is name
of categories which corresponds to data coding order. Polytomous DIF variable
should have element as NULL. Default is NULL.}

\item{n_cov}{Number of covariates before responses. Default is NULL.}

\item{pid}{Name of candidates' ID variable. Default is NULL.}

\item{keyDfLst}{List of dataframe of 'Item', 'Key', and 'Max_score'
(add Key2 if double key). Default is NULL.}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two
testVec. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference
between two testVec. Default is 4.}

\item{desig_effect}{Value to adjust errors. Default is 1.}

\item{step}{TRUE if DIF analysis is performed on step parameters. Default
is FALSE.}

\item{resltReady}{TRUE if results are ready. Default is FALSE.}

\item{iter}{TRUE to iteratively remove DIF items. Default is TRUE.}

\item{test3term}{Vector of names of testVec to include a three-way interaction
(item, step, DIF variable). Default is NULL.}
}
\description{
This function performs DIF analysis on variables and responses in the data.
Data should be processed to have DIF variables of only categories needed.
This is associated with test named 'test'. An Excel file with a summary of
the DIF analysis results and plots will be saved in 'DIF' folder in the
working directory.
}
\examples{
DIF_VarTests()
}
