% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equate.R
\name{Equate}
\alias{Equate}
\title{Equate}
\usage{
Equate(
  df,
  test,
  vars,
  p_cut = 0.05,
  chi_cut = 10,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  sav_results = TRUE,
  desig_effect = 1,
  steps = FALSE
)
}
\arguments{
\item{df}{Dataframe of five variables of 'item' (anchor labels), 'delta.x', 'error.x', 'delta.y', and 'error.y'. '.x' and '.y' should correspond to order of test names in 'vars'.}

\item{test}{Name of test such as 'AGAT'.}

\item{vars}{Vector of length 2 such as c('VIC','NSW'). Its order corresponds to two tests associated with .x and .y.}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{chi_cut}{Threshold of chi-square difference between two tests. Default is 10.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference between two tests. Default is 4.}

\item{sav_results}{TRUE if an Excel file with chi-square test results and a plot are desired. Default is TRUE.}

\item{desig_effect}{Value to adjust errors. Default is 1.}

\item{steps}{TRUE if DIF analysis is performed on step parameters. Default is FALSE.}
}
\value{
Dataframe of chi-square test results for anchors between two tests.
}
\description{
This function performs chi-square tests (DIF analysis) on a group of anchor items' difference of delta estimates in two tests. An Excel file with test results and flags can be saved in 'equating' folder. Also, one plot is saved in subfolder 'plot' inside 'equating' folder.
}
\examples{
Equate()
}
