% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate2Type.R
\name{equate2Type}
\alias{equate2Type}
\title{equate2Type}
\usage{
equate2Type(
  type = c("Hrz", "Vrt"),
  test,
  grdIntVec = c(2:10),
  forms = c("A", "B"),
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_adj_cut = 4,
  step = FALSE,
  iter = TRUE
)
}
\arguments{
\item{type}{Type of equating. One of c('Hrz', 'Vrt').}

\item{test}{Name of test.}

\item{grdIntVec}{Vector of all grades Default is c(2:10).}

\item{forms}{Forms used in a test. Default is c('A','B')}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two
tests. Default is 0.5.}

\item{DIF_adj_cut}{Threshold of an item's adjusted delta estimate difference
between two tests. Default is 4.}

\item{step}{TRUE if DIF analysis is performed on step parameters.
Default is FALSE.}

\item{iter}{TRUE to iteratively remove DIF items. Default is TRUE}
}
\value{
List of chi-square test results for anchors between two adjacent grades.
}
\description{
This function performs vertical equating. An Excel file named 'Vrt_xxx.xlsx'
is saved in 'equating' folder, each sheet of which is for two
adjacent grades Also saved are scatterplots named 'Vrt_xxx.pdf'.
}
\examples{
equate2Type('Hrz', 'bang', 3)
}
