% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equate_shw.R
\name{Equate_shw}
\alias{Equate_shw}
\title{Equate_shw}
\usage{
Equate_shw(
  test,
  vars,
  var_name = NULL,
  p_cut = 0.05,
  DIF_cut = 0.5,
  DIF_std_cut = 4,
  sav_results = TRUE,
  step = FALSE,
  iter = FALSE,
  sigma = FALSE,
  run
)
}
\arguments{
\item{test}{Name of test.}

\item{vars}{Vector of length 2 such as c('VIC','NSW'). This should exist in
.shw file (.del also for step anchors) such as 'test_VIC.shw' and 'test_NSW.shw'.
This will also be used in plotting. Order is flexible.}

\item{var_name}{Character to add before vars in plots. Default is NULL.}

\item{p_cut}{p value of chi-square test. Default is 0.05.}

\item{DIF_cut}{Threshold of an item's delta estimate difference between two
tests. Default is 0.5.}

\item{DIF_std_cut}{Threshold of an item's standardized delta estimate difference
between two tests. Default is 10.}

\item{sav_results}{TRUE if an Excel file with chi-square test results and
a plot are desired. Default is TRUE.}

\item{step}{TRUE if DIF analysis is performed on step parameters.
Default is FALSE.}

\item{iter}{TRUE to iteratively remove DIF items. Default is FALSE.}

\item{sigma}{Indicator of how 'delta.y' is scaled. If TRUE, it is scaled
to have same mean and sd as 'delta.x'. If FALSE, it has same mean as 'delta.x'.
Default is FALSE.}

\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}
}
\value{
Dataframe of chi-square test results for anchors between two tests.
}
\description{
This function extracts two tests' delta estimates and errors from
corresponding .shw files in 'output' folder and performs chi-square tests
(DIF analysis) on their anchors' delta estimate difference. An Excel file
with test results and flags can be saved in 'equating' folder. Also, one
plot is saved in subfolder 'plot' inside 'equating' folder.
}
\examples{
Equate_shw(test='elana_math', vars=c('NSW', 'VIC'))
}
