% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiva_tbl.R
\name{equiva_tbl}
\alias{equiva_tbl}
\title{equiva_tbl}
\usage{
equiva_tbl(
  test,
  est_type = "wle",
  slope = NULL,
  intercept = NULL,
  extrapolation = FALSE,
  run
)
}
\arguments{
\item{test}{Name of test.}

\item{est_type}{Type of ability estimate to use for score equivalence table,
'wle' or 'mle'. Default is 'wle'.}

\item{slope}{Slope to multiply ability estimates. Default is NULL}

\item{intercept}{Value/intercept to add to ability estimates. Default is NULL.}

\item{extrapolation}{Whether to extrapolate the minimum and maximum estimates.
Default is FALSE.}

\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}
}
\value{
Dataframe of scaled ability estimates.
}
\description{
This function extracts desired type of ability estimates from compressed
ConQuest .cqs file and generate scaled scores with specified slope and
intercept. This is associated with test named 'test'.
}
\examples{
equiva_tbl(test='Writing')
}
