% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cas.R
\name{est_cas}
\alias{est_cas}
\title{est_cas}
\usage{
est_cas(folder = here::here("output"), test, sav_xlsx = TRUE)
}
\arguments{
\item{folder}{Folder where 'test.cas' file is located. Default is 'output' folder.}

\item{test}{Name of test.}

\item{sav_xlsx}{TRUE if an Excel file is desired. Default is TRUE.}
}
\value{
Dataframe with variables of 'Pid', 'ScoreRaw', 'ScoreMax', 'Est', and 'SE'.
}
\description{
This function extracts ability estimation from 'test.cas' file. This is associated with test named 'test'.
}
\examples{
est_cas(test='HUM_2')
}
