% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_its.R
\name{file_its}
\alias{file_its}
\title{file_its}
\usage{
file_its(folder = NULL, test, DIFVar)
}
\arguments{
\item{folder}{Folder where subgroups' calibration results are located.
Default is the subfolder with name of argument 'DIFVar' within 'DIF'
folder (folder=NULL).}

\item{test}{Name of test.}

\item{DIFVar}{Name of variable to perform DIF analysis on.}
}
\value{
Dataframe of delta, facility, etc..
}
\description{
This function extracts item analysis statistics from CQ output file 'test.its'.
This is associated with test named 'test'.
}
\examples{
# Not run
# file_its(test='AHU', DIFVar='Gender')
}
