% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{fit_stats}
\alias{fit_stats}
\title{fit_stats}
\usage{
fit_stats(folder = here::here("output"), test)
}
\arguments{
\item{folder}{Folder where 'test.shw' file is located. Default is 'output' folder.}

\item{test}{Name of test.}
}
\description{
This function extracts fit statistics from 'test.shw' file. This is 
associated with test named 'test'.
}
\examples{
fit_stats(test='elana_poly_score')
}
