% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_resps_cat.R
\name{freq_resps_cat}
\alias{freq_resps_cat}
\title{freq_resps_cat}
\usage{
freq_resps_cat(resp, wide = FALSE, prop = FALSE)
}
\arguments{
\item{resp}{Dataframe of responses where each column is a vector of responses
to one test item.}

\item{wide}{TRUE if wide form is desired. Default is FALSE. When default,
the output has long format same as frequency table in .txt file output
from ConQuest.}

\item{prop}{TRUE if proportion is desired. Default is FALSE.}
}
\value{
Table of each item's category frequencies.
}
\description{
This function . This is associated with test named 'test'.
}
\examples{
freq_resps_cat(resp, TRUE, TRUE)
}
