% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_stats.R
\name{item_stats}
\alias{item_stats}
\title{item_stats}
\usage{
item_stats(folder = here::here("output"), test)
}
\arguments{
\item{folder}{Folder where 'test.its' file is located.}

\item{test}{Name of test.}
}
\description{
This function extracts item statistics such as item-rest correlation, facility, delta, and case number from 'test.its' file. This is associated with test named 'test'.
}
\examples{
item_stats(test='AHU')
}
