% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_stats.R
\name{item_stats}
\alias{item_stats}
\title{item_stats}
\usage{
item_stats(folder = "outoput", test)
}
\arguments{
\item{folder}{Folder where ConQuest output files are located.}

\item{test}{Name of test. Default is 'output'.}
}
\description{
This function extracts item statistics such as item-rest correlation,
facility, delta, and case number from 'xxx_its.txt' and 'xxx_shw.txt' files
and merge them into one dataframe. This is associated
with test named 'test'.
}
\examples{
a <- item_stats(test='math_35')
}
