% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itn_ls.R
\name{itn_ls}
\alias{itn_ls}
\title{itn_ls}
\usage{
itn_ls(folder = here::here("output"), test)
}
\arguments{
\item{folder}{Folder that contains ConQuest output files associated with 'test'. Default is 'output' folder in working directory.}

\item{test}{Name of test.}
}
\description{
This function extract item analysis results for each item from .itn file and put them into a list. This is associated with test named 'test'.
}
