% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itn_summary.R
\name{itn_summary}
\alias{itn_summary}
\title{itn_summary}
\usage{
itn_summary(
  test,
  easy = 90,
  hard = 10,
  iRst = 0.11,
  fit_w = 1.1,
  fit_uw = 1.2,
  dFallThr = 0.5,
  dRiseThr = 0.1,
  ccc_data,
  iType,
  keys
)
}
\arguments{
\item{test}{Name of test.}

\item{easy}{Threshold to flag easy items. Default is 90 (percent correct).}

\item{hard}{Threshold to flag hard items. Default is 10 (percent correct).}

\item{iRst}{Threshold to flag low item-rest correlation statistics.
Default is 0.11.}

\item{fit_w}{Threshold to flag large weighted item fit statistics.
Default is 1.1.}

\item{fit_uw}{Threshold to flag large unweighted item fit statistics.
Default is 1.2.}

\item{dFallThr}{Ability on last bin above which falling distractor is flagged.
Default is 0.5.}

\item{dRiseThr}{Ability on last bin below which rising distractor is unflagged.
Default is 0.1.}

\item{ccc_data}{Data to draw CCC. One element of list output from Function 'CCC_Vernon'.}

\item{iType}{Dataframe with columns of iNum and itype. One element of list
output from Function 'CCC_Vernon'.}

\item{keys}{Dataframe of 'Item', 'Key', and 'Max_score' (add Key2 if double key).}
}
\value{
Dataframe of item statistics with flags.
}
\description{
This function produces item analysis results with flags based on item
statistics and CCC. This is associated with test named 'test'.
}
\examples{
a <- itn_summary(test='math_35', ccc_data=ccc_data, iType=iType)
}
