% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_item2.R
\name{N_item2}
\alias{N_item2}
\title{N_item2}
\usage{
N_item2(run, test)
}
\arguments{
\item{run}{String that indicates run such as 'pre_review' and 'post_review'.}

\item{test}{Name of test.}
}
\description{
This function calculates number of items in a test based on 'test.its' file.
 This is associated with test named 'test'.
}
\examples{
N_item2(run, test='AACA')
}
