% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priority_3.R
\name{priority_3}
\alias{priority_3}
\title{priority_3}
\usage{
priority_3(folder = here::here("output"), test)
}
\arguments{
\item{folder}{Folder that contains ConQuest output files associated with 
'test'. Default is 'output' folder in working directory.}

\item{test}{Name of test.}
}
\value{
Vector of flags.
}
\description{
This function checks response option point biserial (Pt Bis) values in 
'test.itn' file in 'output' folder and flag items with any distractor of 
positive Pt Bis. This is associated with test named 'test'.
}
\examples{
priority_3(test='FPA')
}
