% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qid_its.R
\name{qid_its}
\alias{qid_its}
\title{qid_its}
\usage{
qid_its(folder, test)
}
\arguments{
\item{folder}{Folder that contains ConQuest output files associated with 'test'.}

\item{test}{Name of test.}
}
\value{
Vector of item orders.
}
\description{
This function generates vector of item orders that appeared in .its file. This is associated with test named 'test'.
}
\examples{
qid_its()
}
