% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2one.R
\name{read2one}
\alias{read2one}
\title{read2one}
\usage{
read2one(
  folder = c("results", "DIF", "equating"),
  tests,
  prefix = NULL,
  file_name = NULL,
  difVarVec = NULL
)
}
\arguments{
\item{folder}{Folder where to-be-read files are located. Must be one of
c('results', 'DIF', 'equating').}

\item{tests}{Vector of test names after '_'.}

\item{prefix}{Common prefix in file names before first '_'. Default is NULL.}

\item{file_name}{Name given to new file, each sheet of which is one test's result.
Default is NULL.}

\item{difVarVec}{Vector of DIF variables used to extract results from folders
named as elements in the vector.
Its order corresponds to the alphabetic/numeric order of DIF variables'
two categories in data.}
}
\description{
This function reads from 'folder' files associated with 'tests'
and puts them into one file.
}
\examples{
# Not run
# read2one('results', c('AACA'), 'itn')
}
