% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_keys.R
\name{section_keys}
\alias{section_keys}
\title{section_keys}
\usage{
section_keys(keyDf)
}
\arguments{
\item{keyDf}{Dataframe of 'Item', 'Key', and 'Max_score'. Can have 'Key2',
'Key3', etc. if multiple keys.}
}
\value{
String of characters used in key section of 'test.cqc' file in
'input' folder.
}
\description{
This function uses keyDf to generate key strings for CQ control file. A test
of multiple choice with one key has one row/string, and otherwise multiple
rows/strings (e.g., double key or polytomous scoring).
}
\examples{
section_keys(cdLst[[1]])
}
